% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ant.R
\name{fmAnt$addCapsule}
\alias{fmAnt$addCapsule}
\title{Adds a virtual shape body to the ant}
\arguments{
\item{shapeTypeID}{the shape type as an integer}

\item{capsule}{a \code{\link{fmCapsule}} to set the shape to.}
}
\description{
Adds \strong{capsule} as a virtual shape of type
\strong{shapeTypeID}. An error will occur if \strong{shapeTypeID} was not
defined for the \code{\link{fmExperiment}}.
}
\seealso{
Other fmAnt methods: 
\code{\link{fmAnt$clearCapsules}},
\code{\link{fmAnt$deleteCapsule}},
\code{\link{fmAnt$deleteValue}},
\code{\link{fmAnt$getValue}},
\code{\link{fmAnt$identifiedAt}},
\code{\link{fmAnt$setValue}},
\code{\link{fmAnt.DisplayState}},
\code{\link{fmAnt}}
}
\concept{fmAnt methods}
