% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ant.R
\name{fmAnt$clearCapsules}
\alias{fmAnt$clearCapsules}
\title{Removes all capsules from the virtual shape definition}
\description{
Removes all capsules from the virtual shape definition
}
\seealso{
Other fmAnt methods: 
\code{\link{fmAnt$addCapsule}},
\code{\link{fmAnt$deleteCapsule}},
\code{\link{fmAnt$deleteValue}},
\code{\link{fmAnt$getValue}},
\code{\link{fmAnt$identifiedAt}},
\code{\link{fmAnt$setValue}},
\code{\link{fmAnt.DisplayState}},
\code{\link{fmAnt}}
}
\concept{fmAnt methods}
