% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ant.R
\name{fmAnt$deleteCapsule}
\alias{fmAnt$deleteCapsule}
\title{Removes a capsule from the list of virtual shape parts.}
\arguments{
\item{index}{the index to remove (starts at 1)}
}
\description{
Removes a capsule from the list of virtual shape parts.
}
\seealso{
Other fmAnt methods: 
\code{\link{fmAnt$addCapsule}},
\code{\link{fmAnt$clearCapsules}},
\code{\link{fmAnt$deleteValue}},
\code{\link{fmAnt$getValue}},
\code{\link{fmAnt$identifiedAt}},
\code{\link{fmAnt$setValue}},
\code{\link{fmAnt.DisplayState}},
\code{\link{fmAnt}}
}
\concept{fmAnt methods}
