% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ant.R
\name{fmAnt$deleteValue}
\alias{fmAnt$deleteValue}
\title{Deletes a user defined metadata}
\arguments{
\item{key}{a character for the wanted key.}

\item{time}{the time of the timed value we want to remove.}
}
\description{
Removes the value sets for \strong{key} at \strong{time}. An
error will occurs if no value were previously set for \strong{key}
at \strong{time}.
}
\seealso{
Other fmAnt methods: 
\code{\link{fmAnt$addCapsule}},
\code{\link{fmAnt$clearCapsules}},
\code{\link{fmAnt$deleteCapsule}},
\code{\link{fmAnt$getValue}},
\code{\link{fmAnt$identifiedAt}},
\code{\link{fmAnt$setValue}},
\code{\link{fmAnt.DisplayState}},
\code{\link{fmAnt}}
}
\concept{fmAnt methods}
