% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ant.R
\name{fmAnt$getValue}
\alias{fmAnt$getValue}
\title{Gets user defined timed metadata.}
\arguments{
\item{key}{a character for the wanted key.}

\item{time}{a \code{\link{fmTime}} we want the value for.}
}
\value{
either a boolean, integer, numerical, character or
\code{\link{fmTime}}.
}
\description{
Gets user defined timed metadata.
}
\seealso{
Other fmAnt methods: 
\code{\link{fmAnt$addCapsule}},
\code{\link{fmAnt$clearCapsules}},
\code{\link{fmAnt$deleteCapsule}},
\code{\link{fmAnt$deleteValue}},
\code{\link{fmAnt$identifiedAt}},
\code{\link{fmAnt$setValue}},
\code{\link{fmAnt.DisplayState}},
\code{\link{fmAnt}}
}
\concept{fmAnt methods}
