% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ant.R
\name{fmAnt$identifiedAt}
\alias{fmAnt$identifiedAt}
\title{Gets the tag ID identifiying this ant at a given time}
\arguments{
\item{time}{a \code{\link{fmTime}} we want an identification for}
}
\value{
an integer which is the identification for this ant at
\strong{time}.
}
\description{
Gets the tag ID identifiying this ant at a given
\strong{time}. An error will occurs if there is no identification at
\strong{time} fort this ant.
}
\seealso{
Other fmAnt methods: 
\code{\link{fmAnt$addCapsule}},
\code{\link{fmAnt$clearCapsules}},
\code{\link{fmAnt$deleteCapsule}},
\code{\link{fmAnt$deleteValue}},
\code{\link{fmAnt$getValue}},
\code{\link{fmAnt$setValue}},
\code{\link{fmAnt.DisplayState}},
\code{\link{fmAnt}}
}
\concept{fmAnt methods}
