% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ant.R
\name{fmAnt$setValue}
\alias{fmAnt$setValue}
\title{Sets user defined timed metadata.}
\arguments{
\item{key}{a character for the wanted key.}

\item{value}{either a boolean, integer, numerical, character or
\code{link{fmTime}} we want to set \strong{key} to.}

\item{time}{a \code{\link{fmTime}} were \strong{key} will become
\strong{value}.}
}
\description{
Sets \strong{key} to \strong{value} starting at \strong{time}. Errors
could occurs if:
\itemize{
\item \strong{key} is not defined in \code{\link{fmExperiment}}.
\item \strong{time} is  \code{\link{fmTimeForever}()}.
\item \strong{value} is not the right type for \strong{key}
}
}
\seealso{
Other fmAnt methods: 
\code{\link{fmAnt$addCapsule}},
\code{\link{fmAnt$clearCapsules}},
\code{\link{fmAnt$deleteCapsule}},
\code{\link{fmAnt$deleteValue}},
\code{\link{fmAnt$getValue}},
\code{\link{fmAnt$identifiedAt}},
\code{\link{fmAnt.DisplayState}},
\code{\link{fmAnt}}
}
\concept{fmAnt methods}
