% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ant.R
\docType{data}
\name{fmAnt.DisplayState}
\alias{fmAnt.DisplayState}
\title{Describes the posible values for \code{\link{fmAnt$displayStatus}}}
\format{
An object of class \code{list} of length 3.
}
\usage{
fmAnt.DisplayState
}
\description{
Describes the posible values for \code{\link{fmAnt$displayStatus}}
}
\seealso{
Other fmAnt methods: 
\code{\link{fmAnt$addCapsule}},
\code{\link{fmAnt$clearCapsules}},
\code{\link{fmAnt$deleteCapsule}},
\code{\link{fmAnt$deleteValue}},
\code{\link{fmAnt$getValue}},
\code{\link{fmAnt$identifiedAt}},
\code{\link{fmAnt$setValue}},
\code{\link{fmAnt}}
}
\concept{fmAnt methods}
\keyword{datasets}
