% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ant.R
\name{fmAnt}
\alias{fmAnt}
\title{The main object of an Experiment}
\description{
Ant are the main object of interest of an
\code{\link{fmExperiment}}. They are identified from tags with
\code{\link{fmIdentification}}, have a virtual shape to perform
collision and interaction detection, and holds user defined
metadata.

Ant can only be created from an
\code{\link{fmExperiment}} with
\code{\link{fmExperiment$createAnt}()}.

Ant are uniquely identified by their $ID. By
convention we use decimal notation with up to two \code{0} prefix
to display these AntID, as returned by
\code{\link{fmFormatAntID}()}.

Instead of working directly with tag ID
\strong{fort-myrmidon} uses \code{\link{fmIdentification}} to relate
a tag value to an Ant. An Ants could have different
identifications, allowing us to use different tag ID to refer
to the same individuals. One would use
\code{\link{fmAnt$identifiedAt}()} to obtain the tag ID that
identifies an Ant at a given \code{\link{fmTime}}.

Each Ant has an associated virtual shape that is used
to compute instantaneous collision detection (
\code{\link{fmQueryCollideFrames}()} ), or timed ant interactions
( \code{\link{fmQueryComputeAntInteractions}()} ). These shapes
can be defined manually in \strong{fort-studio} or programmatically
accessed with $capsules and modified with
\code{\link{fmAnt$addCapsule}()},
\code{\link{fmAnt$deleteCapsule}()} and
\code{\link{fmAnt$clearCapsules}()}.

Basic visualization of Experiment data can be done
through \strong{fort-studio}. Ants are displayed according to their
$displayStatus and $displayColor. $displayStatus should be one
of the values described in \code{\link{fmAnt.DisplayState}}

Ant can stores timed user defined metadata. These are
modifiable using \code{\link{fmAnt$setValue}()} and
\code{\link{fmAnt$deleteValue}()} and accesible through
\code{\link{fmAnt$getValue}()}.
}
\section{Fields}{

\describe{
\item{\code{identifications}}{a read-only list of
\code{\link{fmIdentification}} for this Ant sorted in
\code{\link{fmTime}}}

\item{\code{capsule}}{a read-only list of named tuple
(type:integer,capsule:\code{link{fmCapsule}}) that describes
the virtual shape of the ant.}

\item{\code{ID}}{the unique ID for this Ant}

\item{\code{displayColor}}{the \code{\link{fmColor}} used to display this
ant in \strong{fort-studio}.}

\item{\code{displayStatus}}{the display status of the Ant in
\strong{fort-studio}, should only be one of the value in
\code{\link{fmAnt.DisplayState}}}
}}

\seealso{
Other fmAnt methods: 
\code{\link{fmAnt$addCapsule}},
\code{\link{fmAnt$clearCapsules}},
\code{\link{fmAnt$deleteCapsule}},
\code{\link{fmAnt$deleteValue}},
\code{\link{fmAnt$getValue}},
\code{\link{fmAnt$identifiedAt}},
\code{\link{fmAnt$setValue}},
\code{\link{fmAnt.DisplayState}}
}
\concept{fmAnt methods}
