% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ant.R
\docType{data}
\name{fmAntDisplayState}
\alias{fmAntDisplayState}
\title{Enum list for \code{fort::myrmidon::Ant::DisplayState}}
\format{
An object of class \code{list} of length 3.
}
\usage{
names(which(a$displayState() == fmAntDisplayState))
}
\description{
list that represent the
    \href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Ant/fort::myrmidon::Ant::DisplayState}{fort::myrmidon::Ant:DisplayState}
    enum
}
\keyword{datasets}
