% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Queries.R
\docType{class}
\name{fmAntInteraction-class}
\alias{fmAntInteraction-class}
\title{A S4 class describing an interaction between two ant}
\description{
A S4 class describing an interaction between two ant
}
\section{Fields}{

\describe{
\item{\code{ant1}}{the id of the first ant}

\item{\code{ant2}}{the id of the second ant}

\item{\code{start}}{the start of the interaction as a POSIXct}

\item{\code{end}}{the end of the interaction as a POSIXct}

\item{\code{types}}{a 2 column matrix with the virtual shape ID of the ant
interacting, first column refers to the first ant, and second
column to the second ant.}

\item{\code{ant1Trajectory}}{the trajectory of the first ant during the
interaction, optionally reported.}

\item{\code{ant2Trajectory}}{the trajectory of the second ant during the
interaction, optionally reported.}
}}

