% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experiment.R
\docType{data}
\name{fmAntMetadataType}
\alias{fmAntMetadataType}
\title{Enum list for \code{fort::myrmidon::AntMetadataType}}
\format{
An object of class \code{list} of length 5.
}
\usage{
names(which( mdTypeValue == fmAntMetadataType) )
}
\description{
list that represent the
    \href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon/fort::myrmidon::AntMetadataType}{fort::myrmidon::AntMetadataType}
    enum
}
\keyword{datasets}
