% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntStaticValue.R
\name{fmAntStaticBool}
\alias{fmAntStaticBool}
\title{creates a \code{\link{fmAntStaticValue}} from a boolean}
\arguments{
\item{v}{the boolean to represent}
}
\value{
a \code{\link{fmAntStaticValue}} of type \code{BOOL}, see \code{\link{fmAntMetadataType}}
}
\description{
creates a \code{\link{fmAntStaticValue}} from a boolean
}
