% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntStaticValue.R
\name{fmAntStaticInteger}
\alias{fmAntStaticInteger}
\title{creates a \code{\link{fmAntStaticValue}} from an integer}
\arguments{
\item{v}{the integer to represent}
}
\value{
a \code{\link{fmAntStaticValue}} of type \code{INTEGER}, see \code{\link{fmAntMetadataType}}
}
\description{
creates a \code{\link{fmAntStaticValue}} from an integer
}
