% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntStaticValue.R
\name{fmAntStaticNumeric}
\alias{fmAntStaticNumeric}
\title{creates a \code{\link{fmAntStaticValue}} from a numerical value}
\arguments{
\item{v}{the numeric to represent}
}
\value{
a \code{\link{fmAntStaticValue}} of type \code{DOUBLE}, see \code{\link{fmAntMetadataType}}
}
\description{
creates a \code{\link{fmAntStaticValue}} from a numerical value
}
