% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntStaticValue.R
\name{fmAntStaticString}
\alias{fmAntStaticString}
\title{creates a \code{\link{fmAntStaticValue}} from a character vector}
\arguments{
\item{v}{the character vector to represent}
}
\value{
a \code{\link{fmAntStaticValue}} of type \code{STRING}, see \code{\link{fmAntMetadataType}}
}
\description{
creates a \code{\link{fmAntStaticValue}} from a character vector
}
