% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntStaticValue.R
\name{fmAntStaticValue}
\alias{fmAntStaticValue}
\title{a wrapper for \code{fort::myrmidon::AntStaticValue}}
\description{
\code{fort::myrmidon::AntStaticValue} are using C++17
    a \code{std::variant}. This wrapper aims to emulate the
    functioanility of these objects. since we cannot mimic the use
    of \code{std::get<TYPE>()} function, these R object uses the
    following methods.
    \cr\cr
    To create an fmStaticValue, please use the following function
    to create an object of the appropriate type:
    \itemize{
    \item \code{\link{fmAntStaticBool}}
    \item \code{\link{fmAntStaticInteger}}
    \item \code{\link{fmAntStaticNumeric}}
    \item \code{\link{fmAntStaticString}}
    \item \code{\link{fmAntStaticTime}}
    }
    \cr\cr
    A \code{fmAntStaticValue} as an assigned type, declared by
    \code{\link{fmAntMetadataType}}. It is not possible to convert
    autonatically one type to another (you have to do it manually).
}
\section{Fields}{

\describe{
\item{\code{toBool}}{cast the object to a boolean value, fails if the type
is not \code{BOOL}}

\item{\code{toInt}}{cast the object to an integer value, fails if the type
is not \code{INTEGER}}

\item{\code{toNumeric}}{cast the object to a numeric value, fails if the
type is not \code{DOUBLE}}

\item{\code{toString}}{cast the object to a character value, fails if the
type is not \code{STRING}}

\item{\code{toTime}}{cast the object to a \code{\link{fmTime}} value,
fails if the type is not \code{TIME}}
}}

