% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Queries.R
\docType{class}
\name{fmAntTrajectory-class}
\alias{fmAntTrajectory-class}
\title{A S4 class describing an Ant trajectory}
\description{
A S4 class describing an Ant trajectory
}
\section{Fields}{

\describe{
\item{\code{ant}}{the id of the ant}

\item{\code{space}}{the space this trajectory is happening}

\item{\code{start}}{the start of the trajectory as a POSIXct}

\item{\code{positions}}{a data.frame with the time,X,Y, Angle and optionally the
ZoneID of the Ant. Time are second offsets from start.}
}}

