% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ant.R
\name{fmCAnt}
\alias{fmCAnt}
\title{A reference class that represent a non-modifiable Ant}
\description{
This is a wrapper for the C++ class
    \href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CAnt}{fort::myrmidon::CAnt}
}
\section{Fields}{

\describe{
\item{\code{identifiedAt}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CAnt/fort::myrmidon::CAnt::IdentifiedAt}{fort::myrmidon::CAnt::IdentifiedAt}}

\item{\code{cIdentifications}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CAnt/fort::myrmidon::CAnt::CIdentifications}{fort::myrmidon::CAnt::CIdentifications}}

\item{\code{antID}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CAnt/fort::myrmidon::CAnt::AntID}{fort::myrmidon::CAnt::AntID}}

\item{\code{formattedID}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CAnt/fort::myrmidon::CAnt::FormattedID}{fort::myrmidon::CAnt::FormattedID}}

\item{\code{displayColor}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CAnt/fort::myrmidon::CAnt::DisplayColor}{fort::myrmidon::CAnt::DisplayColor}}

\item{\code{displayStatus}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CAnt/fort::myrmidon::CAnt::DisplayStatus}{fort::myrmidon::CAnt::DisplayStatus}}

\item{\code{getValue}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CAnt/fort::myrmidon::CAnt::GetValue}{fort::myrmidon::CAnt::GetValue}}

\item{\code{capsules}}{wrapper for
\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CAnt/fort::myrmidon::CAnt::Capsules}{fort::myrmidon::CAnt::Capsules}. It
returns a list with two elements, the 'capsules' element, a
list of \linkS4class{fmCapsule} and 'shapeTypes' their
corresponding shape types.}
}}

