% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{fmCapsule}
\alias{fmCapsule}
\title{Represents a capsule}
\description{
A \code{\link{fmShape}} that describe a capsule,
i.e. the region inside and between two circles
}
\section{Fields}{

\describe{
\item{\code{c1}}{the first circle's center as a numeric vector}

\item{\code{c2}}{the second circle's center as a numeric vector}

\item{\code{r1}}{the first circle's radiu}

\item{\code{r2}}{the second circle's radius}
}}

\seealso{
Other fmShape methods and classes: 
\code{\link{fmCircleCreate}()},
\code{\link{fmCircle}},
\code{\link{fmShapeTypes}},
\code{\link{fmShape}}
}
\concept{fmShape methods and classes}
