% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{fmCapsuleCreate}
\alias{fmCapsuleCreate}
\title{Creates a capsule}
\usage{
fmCapsuleCreate(c1 = c(0, 0), c2 = c(1, 1), r1 = 1, r2 = 1)
}
\arguments{
\item{c1}{the first circle's center}

\item{c2}{the second circle's center}

\item{r1}{the first circle's radius}

\item{r2}{the second circle's radius}
}
\value{
a \code{\link{fmCapsule}}
}
\description{
Creates a capsule
}
\seealso{
Other fmShape methods and classes: 
\code{\link{fmCapsule}},
\code{\link{fmCircleCreate}()},
\code{\link{fmCircle}},
\code{\link{fmPolygonCreate}()},
\code{\link{fmPolygon}},
\code{\link{fmShapeTypes}},
\code{\link{fmShape}}
}
\concept{fmShape methods and classes}
