% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experiment.R
\name{fmCExperiment}
\alias{fmCExperiment}
\title{a reference class wrapper for the C++
    \href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CExperiment/fort::myrmidon::CExperiment}{fort::myrmidon::CExperiment}
    class.}
\description{
a reference class wrapper for the C++
    \href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CExperiment/fort::myrmidon::CExperiment}{fort::myrmidon::CExperiment}
    class.
}
\section{Fields}{

\describe{
\item{\code{absoluteFilePath}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CExperiment/fort::myrmidon::CExperiment::AbsoluteFilePath}{fort::myrmidon::CExperiment::AbsoluteFilePath}}

\item{\code{cSpaces}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CExperiment/fort::myrmidon::CExperiment::CSpaces}{fort::myrmidon::CExperiment::CSpaces}}

\item{\code{cAnts}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CExperiment/fort::myrmidon::CExperiment::CAnts}{fort::myrmidon::CExperiment::CAnts}}

\item{\code{freeIdentificationRangeAt}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CExperiment/fort::myrmidon::CExperiment::FreeIdentificationRangeAt}{fort::myrmidon::CExperiment::FreeIdentificationRangeAt}}

\item{\code{name}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CExperiment/fort::myrmidon::CExperiment::Name}{fort::myrmidon::CExperiment::Name}}

\item{\code{comment}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CExperiment/fort::myrmidon::CExperiment::Comment}{fort::myrmidon::CExperiment::Comment}}

\item{\code{author}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CExperiment/fort::myrmidon::CExperiment::Author}{fort::myrmidon::CExperiment::Author}}

\item{\code{family}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CExperiment/fort::myrmidon::CExperiment::Family}{fort::myrmidon::CExperiment::Family}}

\item{\code{defaultTagSize}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CExperiment/fort::myrmidon::CExperiment::DefaultTagSize}{fort::myrmidon::CExperiment::DefaultTagSize}}

\item{\code{threshold}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CExperiment/fort::myrmidon::CExperiment::Threshold}{fort::myrmidon::CExperiment::Threshold}}

\item{\code{measurementTypeNames}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CExperiment/fort::myrmidon::CExperiment::MeasurementTypeNames}{fort::myrmidon::CExperiment::MeasurementTypeNames}}

\item{\code{antShapeTypeNames}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CExperiment/fort::myrmidon::CExperiment::AntShapeTypeNames}{fort::myrmidon::CExperiment::AntShapeTypeNames}}

\item{\code{antMetadataColumns}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CExperiment/fort::myrmidon::CExperiment::AntMetadataColumns}{fort::myrmidon::CExperiment::AntMetadataColumns}}

\item{\code{getDataInformations}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CExperiment/fort::myrmidon::CExperiment::GetDataInformations}{fort::myrmidon::CExperiment::GetDataInformations}}
}}

