% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identification.R
\name{fmCIdentification}
\alias{fmCIdentification}
\title{Reference Class wrapper for C++ \href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CIdentification/fort::myrmidon::CIdentification}{fort::myrmidon::CIdentification}}
\description{
Reference Class wrapper for C++ \href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CIdentification/fort::myrmidon::CIdentification}{fort::myrmidon::CIdentification}
}
\section{Fields}{

\describe{
\item{\code{tagValue}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CIdentification/fort::myrmidon::CIdentification::TagValue}{fort::myrmidon::CIdentification::TagValue}}

\item{\code{targetAntID}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CIdentification/fort::myrmidon::CIdentification::TargetAntID}{fort::myrmidon::CIdentification::TargetAntID}}

\item{\code{start}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CIdentification/fort::myrmidon::CIdentification::Start}{fort::myrmidon::CIdentification::Start}}

\item{\code{end}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CIdentification/fort::myrmidon::CIdentification::End}{fort::myrmidon::CIdentification::End}}

\item{\code{antPosition}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CIdentification/fort::myrmidon::CIdentification::AntPosition}{fort::myrmidon::CIdentification::AntPosition}}

\item{\code{antAngle}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CIdentification/fort::myrmidon::CIdentification::AntAngle}{fort::myrmidon::CIdentification::AntAngle}}

\item{\code{hasUserDefinedAntPose}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CIdentification/fort::myrmidon::CIdentification::HasUserDefinedAntPose}{fort::myrmidon::CIdentification::HasUserDefinedAntPose}}
}}

