% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{fmCircle}
\alias{fmCircle}
\title{Represents a circle}
\description{
A \code{\link{fmShape}} that describe a circle
}
\section{Fields}{

\describe{
\item{\code{center}}{the circle's center as a numeric vector}

\item{\code{radius}}{the circle's radius}
}}

\seealso{
Other fmShape methods and classes: 
\code{\link{fmCapsuleCreate}()},
\code{\link{fmCapsule}},
\code{\link{fmCircleCreate}()},
\code{\link{fmPolygonCreate}()},
\code{\link{fmPolygon}},
\code{\link{fmShapeList$[[<-}},
\code{\link{fmShapeList$[[}},
\code{\link{fmShapeList$as.list}},
\code{\link{fmShapeList$size}},
\code{\link{fmShapeListCreate}()},
\code{\link{fmShapeList}},
\code{\link{fmShapeTypes}},
\code{\link{fmShape}}
}
\concept{fmShape methods and classes}
