% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{fmCircleCreate}
\alias{fmCircleCreate}
\title{Creates a circle}
\usage{
fmCircleCreate(c1 = c(0, 0), c2 = c(1, 1), r1 = 1, r2 = 1)

fmCircleCreate(c1 = c(0, 0), c2 = c(1, 1), r1 = 1, r2 = 1)
}
\arguments{
\item{c1}{the first circle's center}

\item{c2}{the second circle's center}

\item{r1}{the first circle's radius}

\item{r2}{the second circle's radius}

\item{center}{the circle's center}

\item{radius}{the circle's radius}
}
\value{
a \code{\link{fmCircle}}

a \code{\link{fmCapsule}}
}
\description{
Creates a circle

Creates a capsule
}
\seealso{
Other fmShape methods and classes: 
\code{\link{fmCapsule}},
\code{\link{fmCircle}},
\code{\link{fmShapeTypes}},
\code{\link{fmShape}}

Other fmShape methods and classes: 
\code{\link{fmCapsule}},
\code{\link{fmCircle}},
\code{\link{fmShapeTypes}},
\code{\link{fmShape}}
}
\concept{fmShape methods and classes}
