% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Queries.R
\docType{class}
\name{fmCollision-class}
\alias{fmCollision-class}
\title{a S4 class that describes ant collision (instantaneous interaction)}
\description{
describes an instantaneous interaction between two
    ants. We always have \code{ant1} < \code{ant2} to ensure
    uniqueness of the (ant1,ant2) couple.
}
\section{Fields}{

\describe{
\item{\code{ant1}}{the ID of the first ant}

\item{\code{ant2}}{the ID of the second ant}

\item{\code{zone}}{the ID of the zone where the collision happens}

\item{\code{types}}{a 2 column matrix with the virtual shape ID of the ant
interacting, first column refers to the first ant, and second
column to the second ant.}
}}

