% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Queries.R
\docType{class}
\name{fmCollisionFrame-class}
\alias{fmCollisionFrame-class}
\title{a S4 class that describes ant collisions at a time}
\description{
a S4 class that describes ant collisions at a time
}
\section{Fields}{

\describe{
\item{\code{frameTime}}{the time of the frame as a POSIXct}

\item{\code{space}}{the associated \code{\link{fmSpace$spaceId}} with the
frame}

\item{\code{collisions}}{a list of \code{\link{fmCollision}} of the collision
happening in that space and time.}
}}

