% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Color.R
\name{fmColor}
\alias{fmColor}
\title{Reference class for a RGB color}
\description{
this is a wrapper for the C++ class
    \href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon/fort::myrmidon::Color}{fort::myrmidon::Color}. Since
    \code{std::tuple} does not exist in R, it defines their own
    custom methods.
    \cr\cr
    fmColor can be created with \code{\link{fmRGBColor}} or
    \code{\link{fmDefaultPaletteColor}}
}
\section{Fields}{

\describe{
\item{\code{red}}{the red component of the color as an integer between 0 and 255}

\item{\code{green}}{the green component of the color as an integer between 0 and 255}

\item{\code{blue}}{the blue component of the color as an integer between 0 and 255}
}}

