% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Space.R
\name{fmCSpace}
\alias{fmCSpace}
\title{a Reference Class wrapper for
    \href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CSpace/fort::myrmidon::CSpace}{fort::myrmidon::CSpace}}
\description{
a Reference Class wrapper for
    \href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CSpace/fort::myrmidon::CSpace}{fort::myrmidon::CSpace}
}
\section{Fields}{

\describe{
\item{\code{spaceID}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CSpace/fort::myrmidon::CSpace::CSpaceID}{fort::myrmidon::CSpace::CSpaceID}}

\item{\code{name}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CSpace/fort::myrmidon::CSpace::Name}{fort::myrmidon::CSpace::Name}}

\item{\code{cZones}}{returns a \linkS4class{fmCZoneByID} as
\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CSpace/fort::myrmidon::CSpace::CZones}{fort::myrmidon::CSpace::CZones}}

\item{\code{locateMovieFrame}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CSpace/fort::myrmidon::CSpace::LocateMovieFrame}{fort::myrmidon::CSpace::LocateMovieFrame}}
}}

