% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experiment.R
\name{fmCSpaceByID}
\alias{fmCSpaceByID}
\title{a \code{std::map} of \code{\link{fmCSpace}} by their ID}
\description{
a \code{std::map} of \code{\link{fmCSpace}} by their
    ID. It looks like an R list, but the \code{[[} R operator will
    only report a value if the corresponding ID is in the list, and
    fails otherwise.
}
\section{Fields}{

\describe{
\item{\code{as.list}}{returns the map as an actual list (to iterate over)}

\item{\code{size}}{returns the size of the list}
}}

\examples{
\dontrun{
antByID[[42]] # gets the fmCSpace with ID 42. May fail if there are no fmCSpace with ID 42
antByID.as.list # returns the actual R list, but we won't be able to query for a specific ID.
}
}
