% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zone.R
\name{fmCZone}
\alias{fmCZone}
\title{Reference Class wrapper for the c++
    \href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CZone/fort::myrmidon::CZone}
    class}
\description{
Reference Class wrapper for the c++
    \href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CZone/fort::myrmidon::CZone}
    class
}
\section{Fields}{

\describe{
\item{\code{cDefinitions}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CZone/fort::myrmidon::CZone::CDefinitions}{fort::myrmidon::CZone::CDefinitions}}

\item{\code{name}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CZone/fort::myrmidon::CZone::Name}{fort::myrmidon::CZone::Name}}

\item{\code{zoneID}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CZone/fort::myrmidon::CZone::ZoneID}{fort::myrmidon::CZone::ZoneID}}
}}

