% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zone.R
\name{fmCZoneDefinition}
\alias{fmCZoneDefinition}
\title{Reference Class wrapper for the C++
    \href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::XCZoneDefinition/fort::myrmidon::CZoneDefinition}{fort::myrmidon::CZoneDefinition}
    class}
\description{
Reference Class wrapper for the C++
    \href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::XCZoneDefinition/fort::myrmidon::CZoneDefinition}{fort::myrmidon::CZoneDefinition}
    class
}
\section{Fields}{

\describe{
\item{\code{geometry}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CZoneDefinition/fort::myrmidon::CZoneDefinition::Geometry}{fort::myrmidon::CZoneDefinition::Geometry}}

\item{\code{start}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CZoneDefinition/fort::myrmidon::CZoneDefinition::Start}{fort::myrmidon::CZoneDefinition::Start}}

\item{\code{end}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::CZoneDefinition/fort::myrmidon::CZoneDefinition::End}{fort::myrmidon::CZoneDefinition::End}}
}}

