% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fmDefaultPalette}
\alias{fmDefaultPalette}
\title{Returns myrmidon's default collection of colors}
\usage{
fmDefaultPalette()
}
\value{
a list of RGB 8-bit triplets
}
\description{
Returns myrmidon's default collection of colors
}
\seealso{
Other fmColor methods: 
\code{\link{fmColor}},
\code{\link{fmDefaultPaletteColor}()}
}
\concept{fmColor methods}
