% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fmDefaultPaletteColor}
\alias{fmDefaultPaletteColor}
\title{Returns myrmidon' default color for index, wrapping arround}
\usage{
fmDefaultPaletteColor(i)
}
\arguments{
\item{index}{the wanted index}
}
\value{
a list of RGB 8-bit triplets
}
\description{
Returns myrmidon' default color for index, wrapping arround
}
\seealso{
Other fmColor methods: 
\code{\link{fmColor}},
\code{\link{fmDefaultPalette}()}
}
\concept{fmColor methods}
