% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{fmDuration}
\alias{fmDuration}
\title{Represents a duration with nanosecond precision}
\description{
a fmDuration represents a duration with nanosecond
precision
}
\seealso{
Other fmDuration methods: 
\code{\link{'+.Rcpp_fmDuration'}},
\code{\link{'-.Rcpp_fmDuration'}},
\code{\link{'<.Rcpp_fmDuration'}},
\code{\link{'<=.Rcpp_fmDuration'}},
\code{\link{'==.Rcpp_fmDuration'}},
\code{\link{'>.Rcpp_fmDuration'}},
\code{\link{'>=.Rcpp_fmDuration'}},
\code{\link{fmDurationParse}()},
\code{\link{fmHour}()},
\code{\link{fmMicrosecond}()},
\code{\link{fmMillisecond}()},
\code{\link{fmMinute}()},
\code{\link{fmNanosecond}()},
\code{\link{fmSecond}()}
}
\concept{fmDuration methods}
