% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fmDurationParse}
\alias{fmDurationParse}
\title{Parses a string to a fmDuration}
\usage{
fmDurationParse(input)
}
\arguments{
\item{input}{a string in the format \verb{[amount][unit]} as a
duration. Valid units are 'h','m','s','ms','us','ns'. The
pattern can be repeated (i.e. '4m32s' is valid).}
}
\value{
a \code{\link{fmDuration}} that is equal to input
}
\description{
Parses a string to a fmDuration
}
\examples{
fmDurationParse("1m2s")$show() == '1m2s'
}
\seealso{
Other fmDuration methods: 
\code{\link{'+.Rcpp_fmDuration'}},
\code{\link{'-.Rcpp_fmDuration'}},
\code{\link{'<.Rcpp_fmDuration'}},
\code{\link{'<=.Rcpp_fmDuration'}},
\code{\link{'==.Rcpp_fmDuration'}},
\code{\link{'>.Rcpp_fmDuration'}},
\code{\link{'>=.Rcpp_fmDuration'}},
\code{\link{fmDuration}},
\code{\link{fmHour}()},
\code{\link{fmMicrosecond}()},
\code{\link{fmMillisecond}()},
\code{\link{fmMinute}()},
\code{\link{fmNanosecond}()},
\code{\link{fmSecond}()}
}
\concept{fmDuration methods}
