% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{fmExperiment$addIdentification}
\alias{fmExperiment$addIdentification}
\title{Adds an fmIdentication to this fmExperiment}
\arguments{
\item{antID}{the $ID of the \code{\link{fmAnt}} to identify.}

\item{tagID}{the tag value to associate with \strong{antID}.}

\item{start}{the first valid time for this fmIdentification. Could
be \code{\link{fmTimeSinceEver}()}.}

\item{end}{the first invalid time for this fmIdentification. Could
be \code{\link{fmTimeForever}()}.}
}
\value{
the newly created \code{\link{fmIdentification}}
}
\description{
Adds a \code{\link{fmIdentification}} to this
fmExperiment. An error occurs if any time overlap with other
fmIdentification would occurs for \strong{antID} or for \strong{tagID}.
}
\seealso{
Other fmExperiment methods: 
\code{\link{fmExperiment$addTrackingDataDirectory}},
\code{\link{fmExperiment$createAntShapeType}},
\code{\link{fmExperiment$createAnt}},
\code{\link{fmExperiment$createMeasurementType}},
\code{\link{fmExperiment$createSpace}},
\code{\link{fmExperiment$deleteAntShapeType}},
\code{\link{fmExperiment$deleteAnt}},
\code{\link{fmExperiment$deleteIdentification}},
\code{\link{fmExperiment$deleteMeasurementType}},
\code{\link{fmExperiment$deleteMetaDataKey}},
\code{\link{fmExperiment$deleteSpace}},
\code{\link{fmExperiment$ensureAllDataIsLoaded}},
\code{\link{fmExperiment$freeIdentificationRangeAt}},
\code{\link{fmExperiment$identificationsAt}},
\code{\link{fmExperiment$removeTrackingDataDirectory}},
\code{\link{fmExperiment$renameMetaDataKey}},
\code{\link{fmExperiment$save}},
\code{\link{fmExperiment$setAntShapeTypeName}},
\code{\link{fmExperiment$setMeasurementTypeName}},
\code{\link{fmExperiment$setMetaDataKey}},
\code{\link{fmExperiment.HEAD_TAIL_MEASUREMENT_TYPE_ID}},
\code{\link{fmExperimentCreate}()},
\code{\link{fmExperimentOpen}()},
\code{\link{fmExperiment}}
}
\concept{fmExperiment methods}
