% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{fmExperiment$addTrackingDataDirectory}
\alias{fmExperiment$addTrackingDataDirectory}
\title{Associates a tracking data directory with this fmExperiment.}
\arguments{
\item{spaceID}{the $ID of the fmSpace to associate data with}

\item{filepath}{the filepath to the tracking data directory}

\item{fixCorruptedData}{if FALSE, in the event the tracking data is
corrupted an error will occur. If TRUE \strong{fort-myrmidon} will
try to fix the data corruption ( with some permanent data loss)
instead.}
}
\value{
the URI used to designate this directory in this
fmExperiment
}
\description{
Associates a tracking data directory to a
\code{\link{fmSpace}} of this fmExperiment. An error could
occur if either: \itemize{ \item \strong{spaceID} is not valid for
this fmExperiment.  \item \strong{filepath} is not a valid tracking
data directory \item \strong{filepath} data would overlap in Time
with data from another directory in \strong{spaceID} \item
\strong{filepath} is associated with another fmSpace \item
\strong{filepath} is using a tag family that is different from the
other directory associated in this experiment.  }
}
\seealso{
Other fmExperiment methods: 
\code{\link{fmExperiment$addIdentification}},
\code{\link{fmExperiment$createAntShapeType}},
\code{\link{fmExperiment$createAnt}},
\code{\link{fmExperiment$createMeasurementType}},
\code{\link{fmExperiment$createSpace}},
\code{\link{fmExperiment$deleteAntShapeType}},
\code{\link{fmExperiment$deleteAnt}},
\code{\link{fmExperiment$deleteIdentification}},
\code{\link{fmExperiment$deleteMeasurementType}},
\code{\link{fmExperiment$deleteMetaDataKey}},
\code{\link{fmExperiment$deleteSpace}},
\code{\link{fmExperiment$ensureAllDataIsLoaded}},
\code{\link{fmExperiment$freeIdentificationRangeAt}},
\code{\link{fmExperiment$identificationsAt}},
\code{\link{fmExperiment$removeTrackingDataDirectory}},
\code{\link{fmExperiment$renameMetaDataKey}},
\code{\link{fmExperiment$save}},
\code{\link{fmExperiment$setAntShapeTypeName}},
\code{\link{fmExperiment$setMeasurementTypeName}},
\code{\link{fmExperiment$setMetaDataKey}},
\code{\link{fmExperiment.HEAD_TAIL_MEASUREMENT_TYPE_ID}},
\code{\link{fmExperimentCreate}()},
\code{\link{fmExperimentOpen}()},
\code{\link{fmExperiment}}
}
\concept{fmExperiment methods}
