% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{fmExperiment$identificationsAt}
\alias{fmExperiment$identificationsAt}
\title{Obtains the AntID <-> TagID correspondances for a time}
\arguments{
\item{time}{the time to query identification for.}
}
\value{
a list of TagID indexed by AntID.
}
\description{
Obtains the AntID <-> TagID correspondances for a
given \strong{time} as a list. Index of the list are the AntID, and
the value the corresponding TagID. If an AntID does not exist
or is not identified by any TagID, the corresponding value in
the list will be NULL.
}
\seealso{
Other fmExperiment methods: 
\code{\link{fmExperiment$addIdentification}},
\code{\link{fmExperiment$addTrackingDataDirectory}},
\code{\link{fmExperiment$createAntShapeType}},
\code{\link{fmExperiment$createAnt}},
\code{\link{fmExperiment$createMeasurementType}},
\code{\link{fmExperiment$createSpace}},
\code{\link{fmExperiment$deleteAntShapeType}},
\code{\link{fmExperiment$deleteAnt}},
\code{\link{fmExperiment$deleteIdentification}},
\code{\link{fmExperiment$deleteMeasurementType}},
\code{\link{fmExperiment$deleteMetaDataKey}},
\code{\link{fmExperiment$deleteSpace}},
\code{\link{fmExperiment$ensureAllDataIsLoaded}},
\code{\link{fmExperiment$freeIdentificationRangeAt}},
\code{\link{fmExperiment$removeTrackingDataDirectory}},
\code{\link{fmExperiment$renameMetaDataKey}},
\code{\link{fmExperiment$save}},
\code{\link{fmExperiment$setAntShapeTypeName}},
\code{\link{fmExperiment$setMeasurementTypeName}},
\code{\link{fmExperiment$setMetaDataKey}},
\code{\link{fmExperiment.HEAD_TAIL_MEASUREMENT_TYPE_ID}},
\code{\link{fmExperimentCreate}()},
\code{\link{fmExperimentOpen}()},
\code{\link{fmExperiment}}
}
\concept{fmExperiment methods}
