% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{fmExperiment$save}
\alias{fmExperiment$save}
\title{Saves the fmExperiment to the filesystem}
\arguments{
\item{filepath}{the filepath to save the fmExperiment to}
}
\description{
Saves the fmExperiment to the filesystem. Errors
occurs if this method is used to change the directory of the
current fmExperiment.
}
\seealso{
Other fmExperiment methods: 
\code{\link{fmExperiment$addIdentification}},
\code{\link{fmExperiment$addTrackingDataDirectory}},
\code{\link{fmExperiment$createAntShapeType}},
\code{\link{fmExperiment$createAnt}},
\code{\link{fmExperiment$createMeasurementType}},
\code{\link{fmExperiment$createSpace}},
\code{\link{fmExperiment$deleteAntShapeType}},
\code{\link{fmExperiment$deleteAnt}},
\code{\link{fmExperiment$deleteIdentification}},
\code{\link{fmExperiment$deleteMeasurementType}},
\code{\link{fmExperiment$deleteMetaDataKey}},
\code{\link{fmExperiment$deleteSpace}},
\code{\link{fmExperiment$ensureAllDataIsLoaded}},
\code{\link{fmExperiment$freeIdentificationRangeAt}},
\code{\link{fmExperiment$identificationsAt}},
\code{\link{fmExperiment$removeTrackingDataDirectory}},
\code{\link{fmExperiment$renameMetaDataKey}},
\code{\link{fmExperiment$setAntShapeTypeName}},
\code{\link{fmExperiment$setMeasurementTypeName}},
\code{\link{fmExperiment$setMetaDataKey}},
\code{\link{fmExperiment.HEAD_TAIL_MEASUREMENT_TYPE_ID}},
\code{\link{fmExperimentCreate}()},
\code{\link{fmExperimentOpen}()},
\code{\link{fmExperiment}}
}
\concept{fmExperiment methods}
