% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{fmExperiment$setMetaDataKey}
\alias{fmExperiment$setMetaDataKey}
\title{Sets or creates the name, type and default value for an user
defined metadata key.}
\arguments{
\item{key}{the key to define}

\item{defaultValue}{the default value for \strong{key}}
}
\description{
Sets or creates the name, type and default value for
an user defined metadata key. the type is inferred from
\strong{defaultValue} which should either be a logical, integer,
numerical, character or \code{\link{fmTime}} value. An error
could occurs if \strong{key} is already defined, \strong{defaultValue}
would change the type for \strong{key} and at least one fmAnt has a
defined timed data for \strong{key}.
}
\seealso{
Other fmExperiment methods: 
\code{\link{fmExperiment$addIdentification}},
\code{\link{fmExperiment$addTrackingDataDirectory}},
\code{\link{fmExperiment$createAntShapeType}},
\code{\link{fmExperiment$createAnt}},
\code{\link{fmExperiment$createMeasurementType}},
\code{\link{fmExperiment$createSpace}},
\code{\link{fmExperiment$deleteAntShapeType}},
\code{\link{fmExperiment$deleteAnt}},
\code{\link{fmExperiment$deleteIdentification}},
\code{\link{fmExperiment$deleteMeasurementType}},
\code{\link{fmExperiment$deleteMetaDataKey}},
\code{\link{fmExperiment$deleteSpace}},
\code{\link{fmExperiment$freeIdentificationRangeAt}},
\code{\link{fmExperiment$identificationsAt}},
\code{\link{fmExperiment$removeTrackingDataDirectory}},
\code{\link{fmExperiment$renameMetaDataKey}},
\code{\link{fmExperiment$save}},
\code{\link{fmExperiment$setAntShapeTypeName}},
\code{\link{fmExperiment$setMeasurementTypeName}},
\code{\link{fmExperiment.HEAD_TAIL_MEASUREMENT_TYPE_ID}},
\code{\link{fmExperimentCreate}()},
\code{\link{fmExperimentOpen}()},
\code{\link{fmExperiment}}
}
\concept{fmExperiment methods}
