% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{fmExperiment}
\alias{fmExperiment}
\title{API entry point}
\description{
An fmExperiment olds a collection of
\code{\link{fmAnt}}, \code{\link{fmIdentification}},
\code{\link{fmSpace}} and \code{\link{fmZone}} and give access
to the identified tracking data instantaneous collision and
interaction detection through \code{\link{fmQuery}}.

fmExperiment are saved to the filesystem in
\code{.myrmidon} files. One can use \code{\link{fmExperimentOpen}()}
and \code{\link{fmExperiment$save}()} to interact with these
files.

One can use
\code{\link{fmExperiment$addTrackingDataDirectory}()} to link
an fmExperiment with some tracking data, organized by
\strong{fort-leto} in a tracking data directory. This data must be
assigned to a \code{\link{fmSpace}} (previously created with
\code{\link{fmExperiment$createSpace}()}).  fmExperiment saves
relative links to these tracking data directory. These paths
are relative, so one can rename a \code{.myrmidon} file on the
filesystem with \code{\link{fmExperiment$save}()}, but it must
remains in the same directory.

In \strong{fort-myrmidon}, tags are not used
directly. Instead user are required to make \code{\link{fmAnt}}
object (through \code{\link{fmExperiment$createAnt}()}) and use
\code{\link{fmIdentification}} (through
\code{\link{fmExperiment$addIdentification}()}) to relate a tag
value to an \code{\link{fmAnt}}. To perform collision and
interaction detection, users can create for each Ant a virtual
shape, made of a collection of \code{\link{fmCapsule}}. Each
Capsule is assigned an AntShapeTypeID (an integer starting from
1) which must be previously defined using
\code{\link{fmExperiment$createAntShapeType}()}. There is no
predefined AntShapeTypeID.

\strong{fort-studio} allows to make measurement on close-up
of each \code{\link{fmAnt}}. These measurement must be assigned
to a type, created with
\code{\link{fmExperiment$createMeasurementType}()}. There is a
predefined, undeletable MeasurementTypeID:
\code{\link{fmExperiment.HEAD_TAIL_MEASUREMENT_TYPE_ID}}. It is
used to automatically determine \code{$antPosition} and
\code{$antAngle} fields of \code{\link{fmIdentification}} from
\strong{fort-studio} measurements.

Each \code{\link{fmAnt}} can also holds a dictionnary
of key/value pairs. The key name, type and initial value for
each Ant must be defined with
\code{\link{fmExperiment$setMetaDataKey}()}. Through
\code{\link{fmAnt$setValue}()}, individual, timed value can be
assigned to each Ant. There are no predefined keys.
}
\section{Fields}{

\describe{
\item{\code{name}}{a character that holds the fmExperiment name.}

\item{\code{author}}{a character that holds the fmExperiment author.}

\item{\code{comment}}{a character that holds the fmExperiment comment.}

\item{\code{spaces}}{a read-only, list of fmSpace of this fmExperiment.}

\item{\code{ants}}{a read-only, list of fmAnt of this fmExperiment.}

\item{\code{absoluteFilePath}}{a read-only, character holding the
associated filepath of this fmExperiment.}

\item{\code{defaultTagSize}}{the default tag size in mm for
fmIdentification in this fmExperiment.}

\item{\code{mesurementTypeNames}}{a read-only list of the fmExperiment
defined measurement types}

\item{\code{antShapeTypeNames}}{a read-only list of the fmExperiment
defined ant shape types}

\item{\code{metaDataKeys}}{a read-only list of the fmExperiment user
defined metadata keys.}
}}

\seealso{
Other fmExperiment methods: 
\code{\link{fmExperiment$addIdentification}},
\code{\link{fmExperiment$addTrackingDataDirectory}},
\code{\link{fmExperiment$createAntShapeType}},
\code{\link{fmExperiment$createAnt}},
\code{\link{fmExperiment$createMeasurementType}},
\code{\link{fmExperiment$createSpace}},
\code{\link{fmExperiment$deleteAntShapeType}},
\code{\link{fmExperiment$deleteAnt}},
\code{\link{fmExperiment$deleteIdentification}},
\code{\link{fmExperiment$deleteMeasurementType}},
\code{\link{fmExperiment$deleteMetaDataKey}},
\code{\link{fmExperiment$deleteSpace}},
\code{\link{fmExperiment$ensureAllDataIsLoaded}},
\code{\link{fmExperiment$freeIdentificationRangeAt}},
\code{\link{fmExperiment$identificationsAt}},
\code{\link{fmExperiment$removeTrackingDataDirectory}},
\code{\link{fmExperiment$renameMetaDataKey}},
\code{\link{fmExperiment$save}},
\code{\link{fmExperiment$setAntShapeTypeName}},
\code{\link{fmExperiment$setMeasurementTypeName}},
\code{\link{fmExperiment$setMetaDataKey}},
\code{\link{fmExperiment.HEAD_TAIL_MEASUREMENT_TYPE_ID}},
\code{\link{fmExperimentCreate}()},
\code{\link{fmExperimentOpen}()}
}
\concept{fmExperiment methods}
