% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fmExperimentCreate}
\alias{fmExperimentCreate}
\title{Creates an Experiement.}
\usage{
fmExperimentCreate(filepath)
}
\arguments{
\item{filepath}{to use for the experiment.}
}
\value{
a \link{\code{fmExperiment}} with the associated filepath
}
\description{
No file will be created a filepath, but it is
required to determine relative path to the eventual Tracking
Data Directory that will be added to the experiment.
}
\seealso{
Other fmExperiment methods: 
\code{\link{fmExperimentOpen}()}
}
\concept{fmExperiment methods}
