% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experiment.R
\name{fmExperimentDataInfo}
\alias{fmExperimentDataInfo}
\title{a reference class that summarizes data information about a \code{\link{fmExperiment}}}
\description{
wrapper for
    \href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::ExperimentDataInfo/fort::myrmidon::SpaceDataInfo}{fort::myrmidon::ExperimentDataInfo}
}
\section{Fields}{

\describe{
\item{\code{frames}}{the number of frame accessible in the space}

\item{\code{start}}{the first captured \code{\link{fmTime}} in the experiment}

\item{\code{end}}{the last captured \code{\link{fmTime}} in the experiment}

\item{\code{spaces}}{a \code{\link{fmSpaceDataInfoByID}}, a map of all \code{\link{fmSpaceDataInfo}} by their corresponding space \code{fmSpace$spaceID()}}
}}

