% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fmExperimentOpen}
\alias{fmExperimentOpen}
\title{Opens an Experiment}
\usage{
fmExperimentOpen(filepath)
}
\arguments{
\item{filepath}{the path to the '.myrmidon' file}
}
\value{
a \link{\code{fmExperiment}}
}
\description{
Opens an Experiment
}
\seealso{
Other fmExperiment methods: 
\code{\link{fmExperimentCreate}()}
}
\concept{fmExperiment methods}
