% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fmFormatAntID}
\alias{fmFormatAntID}
\title{Formats an AntID}
\usage{
fmFormatAntID(antID)
}
\arguments{
\item{antID}{the AntID to format}
}
\value{
a character with the formatted ID
}
\description{
Formats an AntID
}
