% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fmFormatTagID}
\alias{fmFormatTagID}
\title{Formats a TagID}
\usage{
fmFormatTagID(tagID)
}
\arguments{
\item{tagID}{the TagID to format}
}
\value{
a character with the formatted ID
}
\description{
Formats a TagID
}
