% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fmHour}
\alias{fmHour}
\title{A fmDuration in hours}
\usage{
fmHour(h)
}
\arguments{
\item{h}{an amount of hours}
}
\value{
a \code{\link{fmDuration}} that is equal to \code{h} hours
}
\description{
A fmDuration in hours
}
\examples{
fmHour(1.32)
fmHour(-3.2) # fmDuration can be negative
}
\seealso{
Other fmDuration methods: 
\code{\link{'+.Rcpp_fmDuration'}},
\code{\link{'-.Rcpp_fmDuration'}},
\code{\link{'<.Rcpp_fmDuration'}},
\code{\link{'<=.Rcpp_fmDuration'}},
\code{\link{'==.Rcpp_fmDuration'}},
\code{\link{'>.Rcpp_fmDuration'}},
\code{\link{'>=.Rcpp_fmDuration'}},
\code{\link{fmDurationParse}()},
\code{\link{fmDuration}},
\code{\link{fmMicrosecond}()},
\code{\link{fmMillisecond}()},
\code{\link{fmMinute}()},
\code{\link{fmNanosecond}()},
\code{\link{fmSecond}()}
}
\concept{fmDuration methods}
