% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identification.R
\name{fmIdentification$clearUserDefinedAntPose}
\alias{fmIdentification$clearUserDefinedAntPose}
\title{Removes the user defined ant pose.}
\description{
The Ant pose will be reverted to the
automatically computed one from \strong{fort-studio}
}
\seealso{
Other fmIdentification methods: 
\code{\link{fmIdendification}},
\code{\link{fmIdentification$hasDefaultTagSize}},
\code{\link{fmIdentification$hasUserDefinedAntPose}},
\code{\link{fmIdentification$setUserDefinedAntPose}}
}
\concept{fmIdentification methods}
