% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identification.R
\name{fmIdentification$hasDefaultTagSize}
\alias{fmIdentification$hasDefaultTagSize}
\title{Indicates if the Experiment's default tag size should be used}
\value{
\code{TRUE} when \code{$tagSize == 0.0}
}
\description{
Indicates if the Experiment's default tag size should be used
}
\seealso{
Other fmIdentification methods: 
\code{\link{fmIdentification$clearUserDefinedAntPose}},
\code{\link{fmIdentification$hasUserDefinedAntPose}},
\code{\link{fmIdentification$setUserDefinedAntPose}},
\code{\link{fmIdentification}}
}
\concept{fmIdentification methods}
