% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identification.R
\name{fmIdentification$hasUserDefinedAntPose}
\alias{fmIdentification$hasUserDefinedAntPose}
\title{Indicates if the user overided the computed Ant Pose}
\value{
\code{TRUE} when an ant pose was defined with
\code{\link{fmIdentification$setUserDefinedAntPose}}
}
\description{
Indicates if the user overided the computed Ant Pose
}
\seealso{
Other fmIdentification methods: 
\code{\link{fmIdendification}},
\code{\link{fmIdentification$clearUserDefinedAntPose}},
\code{\link{fmIdentification$hasDefaultTagSize}},
\code{\link{fmIdentification$setUserDefinedAntPose}}
}
\concept{fmIdentification methods}
