% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identification.R
\name{fmIdentification$setUserDefinedAntPose}
\alias{fmIdentification$setUserDefinedAntPose}
\title{Overrides the computed Ant Pose with a user defined one.}
\arguments{
\item{antPosition}{the position of the ant relatively to the tag
center, in the tag reference frame. It should be a numerical
vector of size 2.}

\item{antAngle}{the ant angle relatively to the tag angle in radians.}
}
\description{
Sets an user defined ant position relatively to the
tag, overiding the one computed from manual measurements.

To revert to the automatically computed ones, use
\code{\link{fmIdentification$clearUserDefinedAntPose}}
}
\seealso{
Other fmIdentification methods: 
\code{\link{fmIdendification}},
\code{\link{fmIdentification$clearUserDefinedAntPose}},
\code{\link{fmIdentification$hasDefaultTagSize}},
\code{\link{fmIdentification$hasUserDefinedAntPose}}
}
\concept{fmIdentification methods}
