% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identification.R
\name{fmIdentification}
\alias{fmIdentification}
\title{Relates tag to fmAnt}
\description{
Identifications relates tag to \code{\link{fmAnt}}
with time and geometric data.

Identification can only be created from an
\code{\link{fmExperiment}} with
\code{\link{fmExperiment$addIdentification}()}.

Identifications are bounded in \code{\link{fmTime}} in the
range [ $start , $end [. These attributes can respectively be
set to \code{\link{fmTimeSinceEver}()} and
\code{\link{fmTimeForever}()}. Internally \strong{fort-myrmidon}
ensure the validity of all Identifications. It means that:
\itemize{
\item Two Identifications using the same $tagValue cannot overlap in Time.
\item Two Identifications using targeting the same \code{\link{fmAnt}} cannot overlap in Time.
}
If any modifications through
\code{\link{fmExperiment$addIdentification}()}, $start or $end
would violate one of these conditions, an error will be raised.

Identifications also contains geometric informations
on how the detected tag is related to the observed
\code{\link{fmAnt}}. These are the translation and rotation of
the Ant, expressed in the tag coordinate reference. Usually,
this information is automatically generated from the manual
measurement of type
\code{\link{fmExperiment.HEAD_TAIL_MEASUREMENT_TYPE_ID}} made in
\strong{fort-studio}. Alternatively, users can override this
behavior by setting themselves this pose using
\code{\link{fmIdentification$setUserDefinedAntPose}()}. \code{\link{fmIdentification$clearUserDefinedAntPose}()}
can be used to revert to the internally computed pose.

Identifications also contains the information of the
physical tag size used to identify the individual. It can be
accessed and set with $tagSize and. The value \code{0.0}
indicates that the \code{$defaultTagSize} field of
\code{\link{fmExperiment}} should be used. Therefore, for most
\code{\link{fmAnt}}, this field should be kept to \code{0.0},
excepted for a few individuals, for examples, Queens.
}
\section{Fields}{

\describe{
\item{\code{tagValue}}{the tag ID used for this identication}

\item{\code{targetAntID}}{the ant ID targetted by this identification}

\item{\code{start}}{the first valid \code{\link{fmTime}} for this
identification}

\item{\code{end}}{the first invalid \code{\link{fmTime}} for this
identification}

\item{\code{antPosition}}{the relative position of the ant in the tag
reference}

\item{\code{antAngle}}{the relative angle of the ant to the tag}

\item{\code{tagSize}}{the physical tag size, if equal to \code{0.0} the
\code{$defaultTagSize} field of \code{\link{fmExperiment}} will
be used}
}}

\note{
Any angle is measured in radians, with a standard
mathematical convention. Since in image processing the y-axis
is pointing from the top of the image to the bottom, positive
angles appears clockwise.
}
\seealso{
Other fmIdentification methods: 
\code{\link{fmIdentification$clearUserDefinedAntPose}},
\code{\link{fmIdentification$hasDefaultTagSize}},
\code{\link{fmIdentification$hasUserDefinedAntPose}},
\code{\link{fmIdentification$setUserDefinedAntPose}}
}
\concept{fmIdentification methods}
