% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Queries.R
\docType{class}
\name{fmIdentifiedFrame-class}
\alias{fmIdentifiedFrame-class}
\title{a S4 class that describes ant position}
\description{
a S4 class that describes ant position
}
\section{Fields}{

\describe{
\item{\code{frameTime}}{the time of the frame as a POSIXct}

\item{\code{width}}{the width of the image}

\item{\code{height}}{the height of the image}

\item{\code{space}}{the associated \code{\link{fmSpace$spaceId}} with the
frame}

\item{\code{data}}{a data.frame with the ID,X,Y Angle and optionally the
ZoneID of the Ant in this frame}
}}

