% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fmMatcherAnd}
\alias{fmMatcherAnd}
\title{Combines fmMatcher in conjunction}
\usage{
fmMatcherAnd(matchers)
}
\arguments{
\item{matchers}{the \code{\link{fmMatcher}} to combine}
}
\value{
a \code{\link{fmMatcher}} that matches when all
\strong{matchers} matches.
}
\description{
Combines fmMatcher in conjunction
}
\seealso{
Other fmMatcher methods: 
\code{\link{fmMatcherAntAngleGreaterThan}()},
\code{\link{fmMatcherAntAngleSmallerThan}()},
\code{\link{fmMatcherAntDisplacement}()},
\code{\link{fmMatcherAntDistanceGreaterThan}()},
\code{\link{fmMatcherAntDistanceSmallerThan}()},
\code{\link{fmMatcherAntID}()},
\code{\link{fmMatcherAntMetaData}()},
\code{\link{fmMatcherAny}()},
\code{\link{fmMatcherInteractionType}()},
\code{\link{fmMatcherOr}()},
\code{\link{fmMatcher}}
}
\concept{fmMatcher methods}
