% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fmMatcherAntAngleGreaterThan}
\alias{fmMatcherAntAngleGreaterThan}
\title{A fmMatcher that matches ants angles}
\usage{
fmMatcherAntAngleGreaterThan(angle)
}
\arguments{
\item{angle}{the angle in radians to be larger}
}
\value{
a \code{\link{fmMatcher}} that matches when the absolute
angle between two Ant is larger than \strong{angle}.
}
\description{
A fmMatcher that matches ants angles
}
\details{
In the case of trajectories, it matches anything.
}
\seealso{
Other fmMatcher methods: 
\code{\link{fmMatcherAnd}()},
\code{\link{fmMatcherAntAngleSmallerThan}()},
\code{\link{fmMatcherAntDisplacement}()},
\code{\link{fmMatcherAntDistanceGreaterThan}()},
\code{\link{fmMatcherAntDistanceSmallerThan}()},
\code{\link{fmMatcherAntID}()},
\code{\link{fmMatcherAntMetaData}()},
\code{\link{fmMatcherAny}()},
\code{\link{fmMatcherInteractionType}()},
\code{\link{fmMatcherOr}()},
\code{\link{fmMatcher}}
}
\concept{fmMatcher methods}
