% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fmMatcherAntDisplacement}
\alias{fmMatcherAntDisplacement}
\title{A fmMatcher that matches ant displacement.}
\usage{
fmMatcherAntDisplacement(under, minimumGap)
}
\arguments{
\item{under}{the maximal allowed displacement in pixel}

\item{minimumGap}{minimal time Gap between check, as a
\code{\link{fmDuration}}. Can possibly be 0.}
}
\value{
a \code{\link{fmMatcher}} that rejects large displacement
in a tracking gap.
}
\description{
Matches trajectories and interactions where the Ant
displacement between two consecutive position is smaller than
\strong{under}. If \strong{minimumGap} is not zero, this check will be
enforced only if there was at least a duration of \strong{minimumGap}
between two consecutive positions.
}
\seealso{
Other fmMatcher methods: 
\code{\link{fmMatcherAnd}()},
\code{\link{fmMatcherAntAngleGreaterThan}()},
\code{\link{fmMatcherAntAngleSmallerThan}()},
\code{\link{fmMatcherAntDistanceGreaterThan}()},
\code{\link{fmMatcherAntDistanceSmallerThan}()},
\code{\link{fmMatcherAntID}()},
\code{\link{fmMatcherAntMetaData}()},
\code{\link{fmMatcherAny}()},
\code{\link{fmMatcherInteractionType}()},
\code{\link{fmMatcherOr}()},
\code{\link{fmMatcher}}
}
\concept{fmMatcher methods}
