% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fmMatcherAntID}
\alias{fmMatcherAntID}
\title{A fmMatcher that matches AntID}
\usage{
fmMatcherAntID(antID)
}
\arguments{
\item{antID}{the Ant ID to match}
}
\value{
a \code{\link{fmMatcher}} that matches when one of the
considered Ant is \strong{antID}
}
\description{
A fmMatcher that matches AntID
}
\details{
In the case of interactions, matches interaction with one
of the Ant having \strong{antID}.
}
\seealso{
Other fmMatcher methods: 
\code{\link{fmMatcherAnd}()},
\code{\link{fmMatcherAntAngleGreaterThan}()},
\code{\link{fmMatcherAntAngleSmallerThan}()},
\code{\link{fmMatcherAntDisplacement}()},
\code{\link{fmMatcherAntDistanceGreaterThan}()},
\code{\link{fmMatcherAntDistanceSmallerThan}()},
\code{\link{fmMatcherAntMetaData}()},
\code{\link{fmMatcherAny}()},
\code{\link{fmMatcherInteractionType}()},
\code{\link{fmMatcherOr}()},
\code{\link{fmMatcher}}
}
\concept{fmMatcher methods}
