% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fmMatcherAntMetaData}
\alias{fmMatcherAntMetaData}
\title{A fmMatcher that matches metadata key/value}
\usage{
fmMatcherAntMetaData(key, value)
}
\arguments{
\item{key}{(a character) the key to match}

\item{value}{the value to match. Should either be a logical,
integer, numerical, character or \code{\link{fmTime}}.}
}
\value{
a \code{\link{fmMatcher}} that matches when one of the
considered Ant is has \strong{key} equal to \strong{value}.
}
\description{
A fmMatcher that matches metadata key/value
}
\details{
In the case of interactions, matches interaction with one
of the Ant meeting the criterion.
}
\seealso{
Other fmMatcher methods: 
\code{\link{fmMatcherAnd}()},
\code{\link{fmMatcherAntAngleGreaterThan}()},
\code{\link{fmMatcherAntAngleSmallerThan}()},
\code{\link{fmMatcherAntDisplacement}()},
\code{\link{fmMatcherAntDistanceGreaterThan}()},
\code{\link{fmMatcherAntDistanceSmallerThan}()},
\code{\link{fmMatcherAntID}()},
\code{\link{fmMatcherAny}()},
\code{\link{fmMatcherInteractionType}()},
\code{\link{fmMatcherOr}()},
\code{\link{fmMatcher}}
}
\concept{fmMatcher methods}
