% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fmMatcherAny}
\alias{fmMatcherAny}
\title{A fmMatcher that matches anything}
\usage{
fmMatcherAny()
}
\value{
a \code{\link{fmMatcher}} that matches everything.
}
\description{
A fmMatcher that matches anything
}
\seealso{
Other fmMatcher methods: 
\code{\link{fmMatcherAnd}()},
\code{\link{fmMatcherAntAngleGreaterThan}()},
\code{\link{fmMatcherAntAngleSmallerThan}()},
\code{\link{fmMatcherAntDisplacement}()},
\code{\link{fmMatcherAntDistanceGreaterThan}()},
\code{\link{fmMatcherAntDistanceSmallerThan}()},
\code{\link{fmMatcherAntID}()},
\code{\link{fmMatcherAntMetaData}()},
\code{\link{fmMatcherInteractionType}()},
\code{\link{fmMatcherOr}()},
\code{\link{fmMatcher}}
}
\concept{fmMatcher methods}
