% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fmMatcherInteractionType}
\alias{fmMatcherInteractionType}
\title{A fmMatcher that matches interaction types}
\usage{
fmMatcherInteractionType(type1, type2)
}
\arguments{
\item{type1}{(integer) the first ant shape type to match}

\item{type2}{(integer) the second ant shape type to match}
}
\value{
a \code{\link{fmMatcher}} that matches
(\strong{type1},\strong{type2}) and (\strong{type2},\strong{type1}) interactions.
}
\description{
A fmMatcher that matches interaction types
}
\details{
In the case of trajectories, it matches anything.
}
\seealso{
Other fmMatcher methods: 
\code{\link{fmMatcherAnd}()},
\code{\link{fmMatcherAntAngleGreaterThan}()},
\code{\link{fmMatcherAntAngleSmallerThan}()},
\code{\link{fmMatcherAntDisplacement}()},
\code{\link{fmMatcherAntDistanceGreaterThan}()},
\code{\link{fmMatcherAntDistanceSmallerThan}()},
\code{\link{fmMatcherAntID}()},
\code{\link{fmMatcherAntMetaData}()},
\code{\link{fmMatcherAny}()},
\code{\link{fmMatcherOr}()},
\code{\link{fmMatcher}}
}
\concept{fmMatcher methods}
