% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fmMatcherOr}
\alias{fmMatcherOr}
\title{Combines fmMatcher in disjunction}
\usage{
fmMatcherOr(matchers)
}
\arguments{
\item{matchers}{the \code{\link{fmMatcher}} to combine}
}
\value{
a \code{\link{fmMatcher}} that matches when any of the
\strong{matchers} matches.
}
\description{
Combines fmMatcher in disjunction
}
\seealso{
Other fmMatcher methods: 
\code{\link{fmMatcherAnd}()},
\code{\link{fmMatcherAntAngleGreaterThan}()},
\code{\link{fmMatcherAntAngleSmallerThan}()},
\code{\link{fmMatcherAntDisplacement}()},
\code{\link{fmMatcherAntDistanceGreaterThan}()},
\code{\link{fmMatcherAntDistanceSmallerThan}()},
\code{\link{fmMatcherAntID}()},
\code{\link{fmMatcherAntMetaData}()},
\code{\link{fmMatcherAny}()},
\code{\link{fmMatcherInteractionType}()},
\code{\link{fmMatcher}}
}
\concept{fmMatcher methods}
