% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fmMicrosecond}
\alias{fmMicrosecond}
\title{A fmDuration in microseconds}
\usage{
fmMicrosecond(us)
}
\arguments{
\item{us}{an amount of microseconds}
}
\value{
a \code{\link{fmDuration}} that is equal to \code{us} microsecond
}
\description{
A fmDuration in microseconds
}
\examples{
fmMicrosecond(1.32)
fmMicrosecond(-3.2) # fmDuration can be negative
}
\seealso{
Other fmDuration methods: 
\code{\link{'+.Rcpp_fmDuration'}},
\code{\link{'-.Rcpp_fmDuration'}},
\code{\link{'<.Rcpp_fmDuration'}},
\code{\link{'<=.Rcpp_fmDuration'}},
\code{\link{'==.Rcpp_fmDuration'}},
\code{\link{'>.Rcpp_fmDuration'}},
\code{\link{'>=.Rcpp_fmDuration'}},
\code{\link{fmDurationParse}()},
\code{\link{fmDuration}},
\code{\link{fmHour}()},
\code{\link{fmMillisecond}()},
\code{\link{fmMinute}()},
\code{\link{fmNanosecond}()},
\code{\link{fmSecond}()}
}
\concept{fmDuration methods}
