% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fmMillisecond}
\alias{fmMillisecond}
\title{A fmDuration in milliseconds}
\usage{
fmMillisecond(ms)
}
\arguments{
\item{ms}{an amount of milliseconds}
}
\value{
a \code{\link{fmDuration}} that is equal to \code{ms} milliseconds
}
\description{
A fmDuration in milliseconds
}
\examples{
fmMillisecond(1.32)
fmMillisecond(-3.2) # fmDuration can be negative
}
\seealso{
Other fmDuration methods: 
\code{\link{'+.Rcpp_fmDuration'}},
\code{\link{'-.Rcpp_fmDuration'}},
\code{\link{'<.Rcpp_fmDuration'}},
\code{\link{'<=.Rcpp_fmDuration'}},
\code{\link{'==.Rcpp_fmDuration'}},
\code{\link{'>.Rcpp_fmDuration'}},
\code{\link{'>=.Rcpp_fmDuration'}},
\code{\link{fmDurationParse}()},
\code{\link{fmDuration}},
\code{\link{fmHour}()},
\code{\link{fmMicrosecond}()},
\code{\link{fmMinute}()},
\code{\link{fmNanosecond}()},
\code{\link{fmSecond}()}
}
\concept{fmDuration methods}
