% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fmMinute}
\alias{fmMinute}
\title{A fmDuration in minutes}
\usage{
fmMinute(m)
}
\arguments{
\item{m}{an amount of minutes}
}
\value{
a \code{\link{fmDuration}} that is equal to \code{m} minutes
}
\description{
A fmDuration in minutes
}
\examples{
fmMinute(1.32)
fmMinute(-3.2) # fmDuration can be negative
}
\seealso{
Other fmDuration methods: 
\code{\link{'+.Rcpp_fmDuration'}},
\code{\link{'-.Rcpp_fmDuration'}},
\code{\link{'<.Rcpp_fmDuration'}},
\code{\link{'<=.Rcpp_fmDuration'}},
\code{\link{'==.Rcpp_fmDuration'}},
\code{\link{'>.Rcpp_fmDuration'}},
\code{\link{'>=.Rcpp_fmDuration'}},
\code{\link{fmDurationParse}()},
\code{\link{fmDuration}},
\code{\link{fmHour}()},
\code{\link{fmMicrosecond}()},
\code{\link{fmMillisecond}()},
\code{\link{fmNanosecond}()},
\code{\link{fmSecond}()}
}
\concept{fmDuration methods}
