% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fmNanosecond}
\alias{fmNanosecond}
\title{A fmDuration in nanosecond}
\usage{
fmNanosecond(ns)
}
\arguments{
\item{ns}{an INTEGER amount of nanoseconds}
}
\value{
a \code{\link{fmDuration}} that is equal to \code{ns} nanosecond
}
\description{
the smallest fmDuration representation. can only take integer values
}
\examples{
fmNanosecond(1)
fmNanosecond(-3) # fmDuration can be negative
}
\seealso{
Other fmDuration methods: 
\code{\link{'+.Rcpp_fmDuration'}},
\code{\link{'-.Rcpp_fmDuration'}},
\code{\link{'<.Rcpp_fmDuration'}},
\code{\link{'<=.Rcpp_fmDuration'}},
\code{\link{'==.Rcpp_fmDuration'}},
\code{\link{'>.Rcpp_fmDuration'}},
\code{\link{'>=.Rcpp_fmDuration'}},
\code{\link{fmDurationParse}()},
\code{\link{fmDuration}},
\code{\link{fmHour}()},
\code{\link{fmMicrosecond}()},
\code{\link{fmMillisecond}()},
\code{\link{fmMinute}()},
\code{\link{fmSecond}()}
}
\concept{fmDuration methods}
