% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fmPolygonCreate}
\alias{fmPolygonCreate}
\title{Creates a fmPolygon}
\usage{
fmPolygonCreate(vertices)
}
\arguments{
\item{vertices}{the polygon vertices}
}
\value{
a closed \code{\link{fmPolygon}} with the given vertices
}
\description{
Creates a fmPolygon
}
\seealso{
Other fmShape methods and classes: 
\code{\link{fmCapsuleCreate}()},
\code{\link{fmCapsule}},
\code{\link{fmCircleCreate}()},
\code{\link{fmCircle}},
\code{\link{fmPolygon}},
\code{\link{fmShapeTypes}},
\code{\link{fmShape}}
}
\concept{fmShape methods and classes}
