% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{fmQuery}
\alias{fmQuery}
\title{Computations on tracking data}
\description{
fmQuery are computation on an fmExperiment tracking data.
}
\note{
For very small Experiment in number of Ant, most of the query
operation are IO bounded, and the multithreading overhead will
impact performance by 40-50\% in computation time, as threads
are waiting for data inpout to come. This is the case
for\code{\link{fmQueryIdentifyFrames}()}, and
\code{\link{fmQueryComputeAntTrajectories}()}. When collision
detection is required, multi-threading can largely help,
especially with a large number of individuals. Threading is
controlled the singleThreaded args of multi-threaded queries.
}
\seealso{
Other fmQuery methods: 
\code{\link{fmQueryComputeMeasurementFor}()},
\code{\link{fmQueryComputeTagStatistics}()}
}
\concept{fmQuery methods}
