% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{fmQueryCollideFrames}
\alias{fmQueryCollideFrames}
\title{Finds Ant collisions in tracking frames.}
\usage{
fmQueryCollideFrames(
  experiment,
  start = fmTimeSinceEver(),
  end = fmTimeForever(),
  collisionsIgnoreZones = FALSE,
  showProgress = TRUE,
  singleThreaded = FALSE
)
}
\arguments{
\item{experiment}{the \code{\link{fmExperiment}} to query on}

\item{start}{the first tracking time to consider,}

\item{end}{the last tracking time to consider}

\item{collisionsIgnoreZones}{if TRUE zones won't affect collision
detection. All ants can collide regardless of their current
zone.}

\item{showProgress}{display query progress in the console}

\item{singleThreaded}{enforce the use of a single thread.}
}
\value{
a named list of two elements:\itemize{\item\code{$frame}: a
\code{data.frame} summarizing the time, space and dimensions of
each frame \item\code{$positions} a list of \code{data.frame}
with the position of each ant for that frame. Each element of
this list correspond to the same row in \code{$frames}
\item{$collisions} a \code{data.frame} with all collision
happening between \strong{start} and \strong{end}. The field
\code{$frame_row_index} indicates in which row in \code{frames}
this collision happens.}
}
\description{
Finds collision, (punctual interactions) between fmAnt
in tracking frame based on their virtual shape. Current zone
for each ants will also be computed.
}
\seealso{
Other fmQuery methods: 
\code{\link{fmQueryComputeAntInteractions}()},
\code{\link{fmQueryComputeAntTrajectories}()},
\code{\link{fmQueryComputeMeasurementFor}()},
\code{\link{fmQueryComputeTagStatistics}()},
\code{\link{fmQueryGetDataInformations}()},
\code{\link{fmQueryGetMetaDataKeyRanges}()},
\code{\link{fmQueryGetTagCloseUps}()},
\code{\link{fmQueryIdentifyFrames}()},
\code{\link{fmQuery}}
}
\concept{fmQuery methods}
