% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{fmQueryComputeAntInteractions}
\alias{fmQueryComputeAntInteractions}
\title{Computes ant interactions for ants.}
\usage{
fmQueryComputeAntInteractions(
  experiment,
  start = fmTimeSinceEver(),
  end = fmTimeForever(),
  maximumGap = fmMillisecond(500),
  matcher = fmMatcherAny(),
  collisionsIgnoreZones = FALSE,
  reportFullTrajectories = FALSE,
  segmentOnMatcherValueChange = FALSE,
  showProgress = TRUE,
  singleThreaded = FALSE
)
}
\arguments{
\item{experiment}{the \code{\link{fmExperiment}} to query on}

\item{start}{the first tracking time to consider,}

\item{end}{the last tracking time to consider}

\item{maximumGap}{the maximum gap in tracking before cutting the
trajectory or interactions in two different object. Use an
insanely large value ( such as \code{fmHour(24*365)} to disable
cutting of trajectory). Trajectory will always be cut when an
ant moves from a space to another.}

\item{matcher}{a \code{\link{fmMatcher}} to reduce the query to
wanted criterion. You can use \code{\link{fmMatcherAny}()} to
match anything.}

\item{collisionsIgnoreZones}{if TRUE zones won't affect collision
detection. All ants can collide regardless of their current
zone.}

\item{reportFullTrajectories}{if TRUE will report for each
interaction the full sub-trajectory, otherwise just a
summary. see the return value.}

\item{segmentOnMatcherValueChange}{if TRUE, when a combined matcher
( AntMetadata("behavior" == "grooming") ||
AntMetadata("behavior" == "sleeping" ) change values ( goes
from "grooming" to "sleeping" ) it will create a new trajectory
or interaction.}

\item{showProgress}{display query progress in the console}

\item{singleThreaded}{enforce the use of a single thread.}
}
\value{
Depending on \strong{reportFullTrajectories}:\itemize{\item if
\code{FALSE}: reports a single \code{data.frame} with the
interactions, with their start and end date, types, and a
summary of each ant positions and zones.\item \code{TRUE} A
named list of 3 elements: \itemize{\item
\code{$trajectories_summary} the start time and space of each
trajectories.\item\code{$trajectories} timed position of the
ant during the trajectory.\item\code{$interactions}: a
\code{data.frame} which contains the interactions start and end
date, and the indexes of \code{$trajectories} that correspond
to each ant sub-trajectory. I.e. for the first ant
\itemize{\item\code{$ant1.trajectory.row} the corresponding
index in \code{$trajectory}. \item\code{$ant1.trajectory.start}
the row index in the \code{data.frame}
\code{$trajectory[[$ant1.trajectory.row]]} that correspond to
the interaction start time. \item\code{$ant1.trajectory.end}:
likewise, the row index that correspond to the interaction
end}}}
}
\description{
Computes ant interactions, i.e. time intervals where
two Ants collides. These will be reported ordered by ending
time.  The parameter \strong{reportFullTrajectories} controls if
full trajectories should be reported or if only a summary
should be reported. The former have an high impact on the
amount of RAM required to perform the query efficiently or at
all.
}
\seealso{
Other fmQuery methods: 
\code{\link{fmQueryCollideFrames}()},
\code{\link{fmQueryComputeAntTrajectories}()},
\code{\link{fmQueryComputeMeasurementFor}()},
\code{\link{fmQueryComputeTagStatistics}()},
\code{\link{fmQueryGetDataInformations}()},
\code{\link{fmQueryGetMetaDataKeyRanges}()},
\code{\link{fmQueryIdentifyFrames}()},
\code{\link{fmQuery}}
}
\concept{fmQuery methods}
