% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{fmQueryComputeAntTrajectories}
\alias{fmQueryComputeAntTrajectories}
\title{Computes trajectories for ants.}
\usage{
fmQueryComputeAntTrajectories(
  experiment,
  start = fmTimeSinceEver(),
  end = fmTimeForever(),
  maximumGap = fmMillisecond(500),
  matcher = fmMatcherAny(),
  computeZones = FALSE,
  segmentOnMatcherValueChange = FALSE,
  showProgress = TRUE,
  singleThreaded = FALSE
)
}
\arguments{
\item{experiment}{the \code{\link{fmExperiment}} to query on}

\item{start}{the first tracking time to consider,}

\item{end}{the last tracking time to consider}

\item{maximumGap}{the maximum gap in tracking before cutting the
trajectory in two different object. Use an insanely large value
( such as \code{fmHour(24*365)} to disable cutting of
trajectory). Trajectory will always be cut when an ant moves
from a space to another.}

\item{matcher}{a \code{\link{fmMatcher}} to reduce the query to
wanted criterion. You can use \code{\link{fmMatcherAny}()} to
match anything.}

\item{computeZones}{computes the zone of the Ant. Otherwise
\code{0} will always be reported.}

\item{segmentOnMatcherValueChange}{if TRUE, when a combined matcher
( AntMetadata("behavior" == "grooming") ||
AntMetadata("behavior" == "sleeping" ) change values ( goes
from "grooming" to "sleeping" ) it will create a new trajectory
or interaction.}

\item{showProgress}{display query progress in the console}

\item{singleThreaded}{enforce the use of a single thread.}
}
\value{
a named list of two elements:
\itemize{\code{$trajectories_summary}: a \code{data.frame}
summarizing the ant, space and starting time of the
trajectory.\item\code{$trajectories} a list of
\code{data.frame} with the position of the ant during the
trajectory \code{$time} is the offset from the trajectory
\code{$start} time in seconds.}
}
\description{
Computes ant trajectories for the fmExperiment. A
trajectory is the consecutive position of an Ant in a Space,
with detection gap smaller than \strong{maximumGap}. These will be
reported ordered by ending time.
}
\seealso{
Other fmQuery methods: 
\code{\link{fmQueryCollideFrames}()},
\code{\link{fmQueryComputeAntInteractions}()},
\code{\link{fmQueryComputeMeasurementFor}()},
\code{\link{fmQueryComputeTagStatistics}()},
\code{\link{fmQueryGetDataInformations}()},
\code{\link{fmQueryGetMetaDataKeyRanges}()},
\code{\link{fmQueryGetTagCloseUps}()},
\code{\link{fmQueryIdentifyFrames}()},
\code{\link{fmQuery}}
}
\concept{fmQuery methods}
