% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fmQueryComputeMeasurementFor}
\alias{fmQueryComputeMeasurementFor}
\title{Queries manual measurements for an Ant}
\usage{
fmQueryComputeMeasurementFor(experiment, antID, typeID = 1L)
}
\arguments{
\item{experiment}{the \code{\link{fmExperiment}} to query}

\item{antID}{the ant to compute for}

\item{typeID}{the wanted measurement type}
}
\value{
a data.frame with the date of the measurement and
estimated size in millimeters and pixels.
}
\description{
Queries manual measurements for an Ant
}
\seealso{
Other fmQuery methods: 
\code{\link{fmQueryCollideFrames}()},
\code{\link{fmQueryComputeAntInteractions}()},
\code{\link{fmQueryComputeAntTrajectories}()},
\code{\link{fmQueryComputeTagStatistics}()},
\code{\link{fmQueryGetDataInformations}()},
\code{\link{fmQueryGetMetaDataKeyRanges}()},
\code{\link{fmQueryGetTagCloseUps}()},
\code{\link{fmQueryIdentifyFrames}()},
\code{\link{fmQuery}}
}
\concept{fmQuery methods}
