% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fmQueryComputeTagStatistics}
\alias{fmQueryComputeTagStatistics}
\title{Computes tag statistics for an experiment}
\usage{
fmQueryComputeTagStatistics(experiment, fixCorruptedData = FALSE)
}
\arguments{
\item{experiment}{the \code{\link{fmExperiment}} to query}

\item{fixCorruptedData}{if \code{TRUE}, will silently fix any
corrupted data. Could lead to the loss of large chunck of
tracking data. Otherwise will raise an error in case of
data corruption.}
}
\value{
a \code{data.frame} with the detection statistics in
\strong{experiment}
}
\description{
Computes tag statistics for an experiment
}
\seealso{
Other fmQuery methods: 
\code{\link{fmQueryCollideFrames}()},
\code{\link{fmQueryComputeAntInteractions}()},
\code{\link{fmQueryComputeAntTrajectories}()},
\code{\link{fmQueryComputeMeasurementFor}()},
\code{\link{fmQueryGetDataInformations}()},
\code{\link{fmQueryGetMetaDataKeyRanges}()},
\code{\link{fmQueryGetTagCloseUps}()},
\code{\link{fmQueryIdentifyFrames}()},
\code{\link{fmQuery}}
}
\concept{fmQuery methods}
