% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fmQueryGetDataInformations}
\alias{fmQueryGetDataInformations}
\title{Collects tracking data information on the fmExperiment}
\usage{
fmQueryGetDataInformations(experiment)
}
\arguments{
\item{experiment}{the \code{\link{fmExperiment}} to query}
}
\value{
a names list with the following items:
\itemize{\code{$frames}: the total number of
frames.\item\code{$start} the first tracked time in the
experiment. \item\code{$end} the last tracked
time. \item\code{$details}: a \code{data.frame} listing the start,
end and number of frame in each tracking data directory and space of
the experiment.}
}
\description{
Collects tracking data information on the fmExperiment
}
\seealso{
Other fmQuery methods: 
\code{\link{fmQueryCollideFrames}()},
\code{\link{fmQueryComputeAntInteractions}()},
\code{\link{fmQueryComputeAntTrajectories}()},
\code{\link{fmQueryComputeMeasurementFor}()},
\code{\link{fmQueryComputeTagStatistics}()},
\code{\link{fmQueryGetMetaDataKeyRanges}()},
\code{\link{fmQueryIdentifyFrames}()},
\code{\link{fmQuery}}
}
\concept{fmQuery methods}
