% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fmQueryGetMetaDataKeyRanges}
\alias{fmQueryGetMetaDataKeyRanges}
\title{get time ranges where a metadata key has a given value}
\usage{
fmQueryGetMetaDataKeyRanges(experiment, key, value)
}
\arguments{
\item{experiment}{the \code{\link{fmExperiment}} to query}

\item{key}{the key to query for}

\item{value}{the value to test for}
}
\value{
a data.frame with the antID, and the start and end time
where key is equal to value.
}
\description{
get time ranges where a metadata key has a given value
}
\seealso{
Other fmQuery methods: 
\code{\link{fmQueryCollideFrames}()},
\code{\link{fmQueryComputeAntInteractions}()},
\code{\link{fmQueryComputeAntTrajectories}()},
\code{\link{fmQueryComputeMeasurementFor}()},
\code{\link{fmQueryComputeTagStatistics}()},
\code{\link{fmQueryGetDataInformations}()},
\code{\link{fmQueryIdentifyFrames}()},
\code{\link{fmQuery}}
}
\concept{fmQuery methods}
