% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fmQueryGetTagCloseUps}
\alias{fmQueryGetTagCloseUps}
\title{Gets close-ups available in the experiment}
\usage{
fmQueryGetTagCloseUps(experiment, fixCorruptedData = FALSE)
}
\arguments{
\item{experiment}{the \code{\link{fmExperiment}} to query}

\item{fixCorruptedData}{if \code{TRUE}, will silently fix '
corrupted data. Could lead to the loss of a few
close-up. Otherwise an error will be raised.}
}
\value{
a data.frame with the closeup
}
\description{
Gets close-ups available in the experiment
}
\seealso{
Other fmQuery methods: 
\code{\link{fmQueryCollideFrames}()},
\code{\link{fmQueryComputeAntInteractions}()},
\code{\link{fmQueryComputeAntTrajectories}()},
\code{\link{fmQueryComputeMeasurementFor}()},
\code{\link{fmQueryComputeTagStatistics}()},
\code{\link{fmQueryGetDataInformations}()},
\code{\link{fmQueryGetMetaDataKeyRanges}()},
\code{\link{fmQueryIdentifyFrames}()},
\code{\link{fmQuery}}
}
\concept{fmQuery methods}
