% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{fmQueryIdentifyFrames}
\alias{fmQueryIdentifyFrames}
\title{Identifies fmAnt in tracking frame}
\usage{
fmQueryIdentifyFrames(
  experiment,
  start = fmTimeSinceEver(),
  end = fmTimeForever(),
  computeZones = FALSE,
  showProgress = TRUE,
  singleThreaded = FALSE
)
}
\arguments{
\item{experiment}{the \code{\link{fmExperiment}} to query on}

\item{start}{the first tracking time to consider,}

\item{end}{the last tracking time to consider}

\item{computeZones}{computes the zone of the Ant, but do not
perform collision detection. If FALSE, \code{0} will be
reported as a zone}

\item{showProgress}{display query progress in the console}

\item{singleThreaded}{enforce the use of a single thread.}
}
\value{
a list of two element \itemize{\item\code{$frames}: a
\code{data.frame} summarizing the time space and dimension of
each frame.\item\code{$positions}: a list of \code{data.frame}
with the position of each ant for that frame. Each element of
this list correspond to the same index in \code{frames}}
}
\description{
Identifies fmAnt in tracking data frame. I.e. finds
fmAnt based on fmIdentification and apply the tag <-> ant
position corrections.
}
\seealso{
Other fmQuery methods: 
\code{\link{fmQueryCollideFrames}()},
\code{\link{fmQueryComputeAntInteractions}()},
\code{\link{fmQueryComputeAntTrajectories}()},
\code{\link{fmQueryComputeMeasurementFor}()},
\code{\link{fmQueryComputeTagStatistics}()},
\code{\link{fmQueryGetDataInformations}()},
\code{\link{fmQueryGetMetaDataKeyRanges}()},
\code{\link{fmQuery}}
}
\concept{fmQuery methods}
