% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Color.R
\name{fmRGBColor}
\alias{fmRGBColor}
\title{Creates an \code{\link{fmColor}} from its RGB value}
\arguments{
\item{xx}{an integer vectorr of size 3 with the RGB value in this
order}
}
\value{
the wanted \code{\link{fmColor}}
}
\description{
Creates an \code{\link{fmColor}} from its RGB value
}
\examples{
white <- fmRGBColor(c(255,255,255))
pureBlue <- fmRGBColor(c(0,0,255))
}
