% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fmSecond}
\alias{fmSecond}
\title{A fmDuration in seconds}
\usage{
fmSecond(s)
}
\arguments{
\item{s}{an amount of seconds}
}
\value{
a \code{\link{fmDuration}} that is equal to \code{s} seconds
}
\description{
A fmDuration in seconds
}
\examples{
fmSecond(1.32)
fmSecond(-3.2) # fmDuration can be negative
}
\seealso{
Other fmDuration methods: 
\code{\link{'+.Rcpp_fmDuration'}},
\code{\link{'-.Rcpp_fmDuration'}},
\code{\link{'<.Rcpp_fmDuration'}},
\code{\link{'<=.Rcpp_fmDuration'}},
\code{\link{'==.Rcpp_fmDuration'}},
\code{\link{'>.Rcpp_fmDuration'}},
\code{\link{'>=.Rcpp_fmDuration'}},
\code{\link{fmDurationParse}()},
\code{\link{fmDuration}},
\code{\link{fmHour}()},
\code{\link{fmMicrosecond}()},
\code{\link{fmMillisecond}()},
\code{\link{fmMinute}()},
\code{\link{fmNanosecond}()}
}
\concept{fmDuration methods}
