% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{fmShape}
\alias{fmShape}
\title{Represents a 2D shape}
\description{
A fmShape describe a 2D geometric shape
}
\section{Fields}{

\describe{
\item{\code{shapeType}}{the shape type, which would be one of the values in fmShapeTypes}
}}

\seealso{
Other fmShape methods and classes: 
\code{\link{fmCapsuleCreate}()},
\code{\link{fmCapsule}},
\code{\link{fmCircleCreate}()},
\code{\link{fmCircle}},
\code{\link{fmPolygonCreate}()},
\code{\link{fmPolygon}},
\code{\link{fmShapeTypes}}
}
\concept{fmShape methods and classes}
