% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{fmShapeList$as.list}
\alias{fmShapeList$as.list}
\title{Copies the fmShapeList as a native R list}
\value{
a R list of copies of all fmShape in the list
}
\description{
Copies the fmShapeList as a native R list
}
\examples{
l <- fmShapeListCreate(list(fmCircleCreate()))
ll = l$as.list()
}
\seealso{
Other fmShape methods and classes: 
\code{\link{fmCapsuleCreate}()},
\code{\link{fmCapsule}},
\code{\link{fmCircleCreate}()},
\code{\link{fmCircle}},
\code{\link{fmPolygonCreate}()},
\code{\link{fmPolygon}},
\code{\link{fmShapeList$[[<-}},
\code{\link{fmShapeList$[[}},
\code{\link{fmShapeList$size}},
\code{\link{fmShapeListCreate}()},
\code{\link{fmShapeList}},
\code{\link{fmShapeTypes}},
\code{\link{fmShape}}
}
\concept{fmShape methods and classes}
