% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{fmShapeList$size}
\alias{fmShapeList$size}
\title{The size of the fmShapeList}
\value{
the size of the \code{\link{fmShapeList}}
}
\description{
the \code{length} function can also be used.
}
\examples{
l <- fmShapeListCreate()
l$size()
\dontrun{length(l)}
}
\seealso{
Other fmShape methods and classes: 
\code{\link{fmCapsuleCreate}()},
\code{\link{fmCapsule}},
\code{\link{fmCircleCreate}()},
\code{\link{fmCircle}},
\code{\link{fmPolygonCreate}()},
\code{\link{fmPolygon}},
\code{\link{fmShapeList$[[<-}},
\code{\link{fmShapeList$[[}},
\code{\link{fmShapeList$as.list}},
\code{\link{fmShapeListCreate}()},
\code{\link{fmShapeList}},
\code{\link{fmShapeTypes}},
\code{\link{fmShape}}
}
\concept{fmShape methods and classes}
