% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{fmShapeList$[[<-}
\alias{fmShapeList$[[<-}
\title{Access elements from the list}
\arguments{
\item{i}{index of the element to access}

\item{v}{the new shape to set at i}
}
\description{
Access elements from the list
}
\examples{
l <- fmShapeListCreate(list(fmCircleCreate()))
l[[1]] = fmCapsuleCreate()# I wanted a capsule there
}
\seealso{
Other fmShape methods and classes: 
\code{\link{fmCapsuleCreate}()},
\code{\link{fmCapsule}},
\code{\link{fmCircleCreate}()},
\code{\link{fmCircle}},
\code{\link{fmPolygonCreate}()},
\code{\link{fmPolygon}},
\code{\link{fmShapeList$[[}},
\code{\link{fmShapeList$as.list}},
\code{\link{fmShapeList$size}},
\code{\link{fmShapeListCreate}()},
\code{\link{fmShapeList}},
\code{\link{fmShapeTypes}},
\code{\link{fmShape}}
}
\concept{fmShape methods and classes}
