% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{fmShapeList}
\alias{fmShapeList}
\title{A list of heterogenous 2D shapes}
\description{
an object that is similar to a list, but could only
accept fmShape. It is implicitly convertible from a native R
list. It can also be copied with $as.list
}
\seealso{
Other fmShape methods and classes: 
\code{\link{fmCapsuleCreate}()},
\code{\link{fmCapsule}},
\code{\link{fmCircleCreate}()},
\code{\link{fmCircle}},
\code{\link{fmPolygonCreate}()},
\code{\link{fmPolygon}},
\code{\link{fmShapeList$[[<-}},
\code{\link{fmShapeList$[[}},
\code{\link{fmShapeList$as.list}},
\code{\link{fmShapeList$size}},
\code{\link{fmShapeListCreate}()},
\code{\link{fmShapeTypes}},
\code{\link{fmShape}}
}
\concept{fmShape methods and classes}
