% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fmShapeListCreate}
\alias{fmShapeListCreate}
\title{Creates a fmShapeList}
\usage{
fmShapeListCreate(shapes = NULL)
}
\arguments{
\item{shapes}{the shape in the list. should be a list of fmCircle,
fmCapsule or fmPolygon.}
}
\value{
a closed \code{\link{fmShapeList}} with the given shapes
}
\description{
Creates a fmShapeList
}
\seealso{
Other fmShape methods and classes: 
\code{\link{fmCapsuleCreate}()},
\code{\link{fmCapsule}},
\code{\link{fmCircleCreate}()},
\code{\link{fmCircle}},
\code{\link{fmPolygonCreate}()},
\code{\link{fmPolygon}},
\code{\link{fmShapeList$[[<-}},
\code{\link{fmShapeList$[[}},
\code{\link{fmShapeList$as.list}},
\code{\link{fmShapeList$size}},
\code{\link{fmShapeList}},
\code{\link{fmShapeTypes}},
\code{\link{fmShape}}
}
\concept{fmShape methods and classes}
