% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\docType{data}
\name{fmShapeTypes}
\alias{fmShapeTypes}
\title{Describes the possible values for fmShape$shapeType}
\format{
An object of class \code{list} of length 3.
}
\usage{
fmShapeTypes
}
\description{
Describes the possible values for fmShape$shapeType
}
\seealso{
Other fmShape methods and classes: 
\code{\link{fmCapsuleCreate}()},
\code{\link{fmCapsule}},
\code{\link{fmCircleCreate}()},
\code{\link{fmCircle}},
\code{\link{fmPolygonCreate}()},
\code{\link{fmPolygon}},
\code{\link{fmShape}}
}
\concept{fmShape methods and classes}
\keyword{datasets}
