% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.R
\name{fmSpace$createZone}
\alias{fmSpace$createZone}
\title{Creates a fmZone for this fmSpace}
\arguments{
\item{name}{the wanted name for the \link{\code{fmZone}}}
}
\value{
the new \link{\code{fmZone}}
}
\description{
Creates a new \link{\code{fmZone}} for this fmSpace.
}
\seealso{
Other fmSpace methods: 
\code{\link{fmSpace$deleteZone}},
\code{\link{fmSpace$locateMovieFrame}},
\code{\link{fmSpace}}
}
\concept{fmSpace methods}
