% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.R
\name{fmSpace$deleteZone}
\alias{fmSpace$deleteZone}
\title{Removes a fmZone for this fmSpace}
\arguments{
\item{zoneID}{the ID of the \code{\link{fmZone}} to remove}
}
\description{
Removes a \code{\link{fmZone}} from this fmSpace.
}
\seealso{
Other fmSpace methods: 
\code{\link{fmSpace$createZone}},
\code{\link{fmSpace$locateMovieFrame}},
\code{\link{fmSpace}}
}
\concept{fmSpace methods}
