% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.R
\name{fmSpace$locateMovieFrame}
\alias{fmSpace$locateMovieFrame}
\title{Locates a movie file and frame number from a fmTime}
\arguments{
\item{time}{the time we want to query}
}
\value{
a named list of the absolute file path to the file, and a
frame number of the frame that was acquired at or just after
time.
}
\description{
Locates a movie file and frame number from a
\code{\link{fmTime}}.
}
\seealso{
Other fmSpace methods: 
\code{\link{fmSpace$createZone}},
\code{\link{fmSpace$deleteZone}},
\code{\link{fmSpace}}
}
\concept{fmSpace methods}
