% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.R
\name{fmSpace}
\alias{fmSpace}
\title{An homogenous coordinate system for tracking data}
\description{
A fmSpace represent the physical space tracked by one
single camera. Coordinates in one fmSpace cannot be compared
with coordinate from another fmSpace.

fmSpace are uniquely identified with their ID.

fmSpace can only be created from an
\code{\link{fmExperiment}} with
\code{\link{fmExperiment$createSpace}()}

Within a single fmSpace, it could be relevant to
define \code{\link{fmZone}} where interaction between
\code{\link{fmAnt}} could be measured. I.e. Ants in different
fmZones won't report interactions.

\code{\link{fmZone}} are manipulated with
\code{\link{fmSpace$createZone}()} and
\code{\link{fmSpace$deleteZone}()}.
}
\section{Fields}{

\describe{
\item{\code{name}}{the fmSpace name as a character}

\item{\code{ID}}{the fmSpace unique identifier in the fmExperiment.}
}}

\seealso{
Other fmSpace methods: 
\code{\link{fmSpace$createZone}},
\code{\link{fmSpace$deleteZone}},
\code{\link{fmSpace$locateMovieFrame}}
}
\concept{fmSpace methods}
