% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experiment.R
\name{fmSpaceDataInfo}
\alias{fmSpaceDataInfo}
\title{a reference class that summarizes data information about a
    \code{\link{fmSpace}} in the \code{\link{fmExperiment}}}
\description{
wrapper for
    \href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::SpaceDataInfo/fort::myrmidon::SpaceDataInfo}{fort::myrmidon::SpaceDataInfo}
}
\section{Fields}{

\describe{
\item{\code{uri}}{the URI used in FORT studio}

\item{\code{name}}{the user defined name of the space}

\item{\code{frames}}{the number of frame accessible in the space}

\item{\code{start}}{the first captured \code{\link{fmTime}} in the space}

\item{\code{end}}{the last captured \code{\link{fmTime}} in the space}

\item{\code{trackingDataDirectories}}{a list of \code{\link{fmTrackingDataDirectoryInfo}}}
}}

