% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\docType{data}
\name{fmTagFamily}
\alias{fmTagFamily}
\title{Enumerates the possible tag family used in the FORT system.}
\format{
An object of class \code{numeric} of length 11.
}
\usage{
fmTagFamily
}
\description{
Enumerates the possible tag family used in the FORT system.
}
\keyword{datasets}
