% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{fmTime$add}
\alias{fmTime$add}
\title{add a \code{\link{fmDuration}} to this \code{\link{fmTime}}}
\arguments{
\item{duration}{the duration to add}
}
\value{
a new \code{\link{fmTime}} which is duration after this fmTime
}
\description{
add a \code{\link{fmDuration}} to this \code{\link{fmTime}}
}
\examples{
\dontrun{
fmTimeCreate()$add(fmSecond(1)) == fmTimeCreate(1.0) # this is the same
}
}
\seealso{
Other fmTime methods: 
\code{\link{'<.Rcpp_fmTime'}},
\code{\link{'<=.Rcpp_fmTime'}},
\code{\link{'==.Rcpp_fmTime'}},
\code{\link{'>.Rcpp_fmTime'}},
\code{\link{'>=.Rcpp_fmTime'}},
\code{\link{fmTime$after}},
\code{\link{fmTime$before}},
\code{\link{fmTime$equals}},
\code{\link{fmTime$round}},
\code{\link{fmTime$sub}},
\code{\link{fmTimeCreate}()},
\code{\link{fmTimeForever}()},
\code{\link{fmTimeNow}()},
\code{\link{fmTimeParse}()},
\code{\link{fmTimeSinceEver}()},
\code{\link{fmTime}}
}
\concept{fmTime methods}
