% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{fmTime$asPOSIXct}
\alias{fmTime$asPOSIXct}
\title{Converts to a R datetime object}
\value{
a datetime object representing this time this value can be
Inf or -Inf if this isInfinite()
}
\description{
Converts to a R datetime object
}
\examples{
fmTimeNow()$asPOSIXct # equivalent to Sys.time()
}
\seealso{
Other fmTime methods: 
\code{\link{'<.Rcpp_fmTime'}},
\code{\link{'<=.Rcpp_fmTime'}},
\code{\link{'==.Rcpp_fmTime'}},
\code{\link{'>.Rcpp_fmTime'}},
\code{\link{'>=.Rcpp_fmTime'}},
\code{\link{fmTime$add}},
\code{\link{fmTime$after}},
\code{\link{fmTime$before}},
\code{\link{fmTime$equals}},
\code{\link{fmTime$isForever}},
\code{\link{fmTime$isInfinite}},
\code{\link{fmTime$isSinceEver}},
\code{\link{fmTime$round}},
\code{\link{fmTime$sub}},
\code{\link{fmTimeCreate}()},
\code{\link{fmTimeForever}()},
\code{\link{fmTimeNow}()},
\code{\link{fmTimeParse}()},
\code{\link{fmTimeSinceEver}()},
\code{\link{fmTime}}
}
\concept{fmTime methods}
