% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{fmTime$equals}
\alias{fmTime$equals}
\title{tests if the fmTime object is euals to another fmTime}
\arguments{
\item{other}{the other fmTime to test for}
}
\value{
\code{TRUE} if this fmTime is equal to other
}
\description{
tests if the fmTime object is euals to another fmTime
}
\seealso{
Other fmTime methods: 
\code{\link{'<.Rcpp_fmTime'}},
\code{\link{'<=.Rcpp_fmTime'}},
\code{\link{'==.Rcpp_fmTime'}},
\code{\link{'>.Rcpp_fmTime'}},
\code{\link{'>=.Rcpp_fmTime'}},
\code{\link{fmTime$add}},
\code{\link{fmTime$after}},
\code{\link{fmTime$before}},
\code{\link{fmTime$round}},
\code{\link{fmTime$sub}},
\code{\link{fmTimeCreate}()},
\code{\link{fmTimeForever}()},
\code{\link{fmTimeNow}()},
\code{\link{fmTimeParse}()},
\code{\link{fmTimeSinceEver}()},
\code{\link{fmTime}}
}
\concept{fmTime methods}
