% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{fmTime$round}
\alias{fmTime$round}
\title{rounds a time to the nearest duration}
\arguments{
\item{duration}{the \code{\link{fmDuration}} to round to.}
}
\value{
a new \code{\link{fmTime}} which this time rounded to duration
}
\description{
rounds a time to the nearest duration
}
\examples{
fmTimeNow()$round(fmHour(1)) # current time rounded to one hour
}
\seealso{
Other fmTime methods: 
\code{\link{'<.Rcpp_fmTime'}},
\code{\link{'<=.Rcpp_fmTime'}},
\code{\link{'==.Rcpp_fmTime'}},
\code{\link{'>.Rcpp_fmTime'}},
\code{\link{'>=.Rcpp_fmTime'}},
\code{\link{fmTime$add}},
\code{\link{fmTime$after}},
\code{\link{fmTime$before}},
\code{\link{fmTime$equals}},
\code{\link{fmTime$sub}},
\code{\link{fmTimeCreate}()},
\code{\link{fmTimeForever}()},
\code{\link{fmTimeNow}()},
\code{\link{fmTimeParse}()},
\code{\link{fmTimeSinceEver}()},
\code{\link{fmTime}}
}
\concept{fmTime methods}
