% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{fmTime$sub}
\alias{fmTime$sub}
\title{computes the time difference between two times}
\arguments{
\item{other}{the other time to substract}
}
\value{
a \code{\link{fmDuration}} which the difference between this fmTime and other
}
\description{
computes the time difference between two times
}
\examples{
start <- fmTimeNow()
# do some computation
ellapsed <- fmTimeNow()$sub(start)
}
\seealso{
Other fmTime methods: 
\code{\link{'<.Rcpp_fmTime'}},
\code{\link{'<=.Rcpp_fmTime'}},
\code{\link{'==.Rcpp_fmTime'}},
\code{\link{'>.Rcpp_fmTime'}},
\code{\link{'>=.Rcpp_fmTime'}},
\code{\link{fmTime$add}},
\code{\link{fmTime$after}},
\code{\link{fmTime$before}},
\code{\link{fmTime$equals}},
\code{\link{fmTime$round}},
\code{\link{fmTimeCreate}()},
\code{\link{fmTimeForever}()},
\code{\link{fmTimeNow}()},
\code{\link{fmTimeParse}()},
\code{\link{fmTimeSinceEver}()},
\code{\link{fmTime}}
}
\concept{fmTime methods}
