% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{fmTime}
\alias{fmTime}
\title{Represents a time}
\description{
a fmTime represents a Time issued by the tracking
system. These times are more precise than the natural ctPOSIX
object used by R. They can also describe the infinite time
\code{\link{fmTimeSinceEver}} and \code{\link{fmTimeForever}}
}
\seealso{
Other fmTime methods: 
\code{\link{'<.Rcpp_fmTime'}},
\code{\link{'<=.Rcpp_fmTime'}},
\code{\link{'==.Rcpp_fmTime'}},
\code{\link{'>.Rcpp_fmTime'}},
\code{\link{'>=.Rcpp_fmTime'}},
\code{\link{fmTime$add}},
\code{\link{fmTime$after}},
\code{\link{fmTime$asPOSIXct}},
\code{\link{fmTime$before}},
\code{\link{fmTime$equals}},
\code{\link{fmTime$isForever}},
\code{\link{fmTime$isInfinite}},
\code{\link{fmTime$isSinceEver}},
\code{\link{fmTime$round}},
\code{\link{fmTime$sub}},
\code{\link{fmTimeCreate}()},
\code{\link{fmTimeForever}()},
\code{\link{fmTimeNow}()},
\code{\link{fmTimeParse}()},
\code{\link{fmTimeSinceEver}()}
}
\concept{fmTime methods}
