% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time.R
\name{fmTimeCPtr}
\alias{fmTimeCPtr}
\title{a Reference Class that represents a pointer to a fmTime}
\description{
in myrmiodn, many methods or static methods takes as
    argument a \code{fort::myrmidon::Time::ConstPtr} a pointer to a
    Time object. C++ \code{nullptr} for these methods represents
    +/-∞. R does not differentiate object and pointer, so this
    class is used to represent a pointer to an
    \linkS4class{fmTime}. use \code{\link{get}} method to get the
    underlying \linkS4class{fmTime} object.
}
\section{Fields}{

\describe{
\item{\code{get}}{return the underlying \linkS4class{fmTime} object.}
}}

