% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time.R
\name{fmTimeCPtrFromAnySEXP}
\alias{fmTimeCPtrFromAnySEXP}
\title{Converts several object type to a \linkS4class{fmTimeCPtr}}
\arguments{
\item{exp}{an object that could be:
\itemize{
\item NULL to represent +/-∞
\item a POSIXct/POSIXlt object from R
\item a fmTimeCPtr, such as one returned by many myrmidon methods
\item a fmTime such as outpute by \code{\link{fmTimeNow}} or \code{\link{fmTimeParse}}
}}
}
\value{
a \linkS4class{fmTimeCPtr} representing the wanted time.
}
\description{
Converts several object type to a \linkS4class{fmTimeCPtr}
}
