% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fmTimeNow}
\alias{fmTimeNow}
\title{Returns current time as a fmTime}
\usage{
fmTimeNow()
}
\value{
the current time as a \code{\link{fmTime}}
}
\description{
Returns current time as a fmTime
}
\seealso{
Other fmTime methods: 
\code{\link{'<.Rcpp_fmTime'}},
\code{\link{'<=.Rcpp_fmTime'}},
\code{\link{'==.Rcpp_fmTime'}},
\code{\link{'>.Rcpp_fmTime'}},
\code{\link{'>=.Rcpp_fmTime'}},
\code{\link{fmTime$add}},
\code{\link{fmTime$after}},
\code{\link{fmTime$asPOSIXct}},
\code{\link{fmTime$before}},
\code{\link{fmTime$equals}},
\code{\link{fmTime$isForever}},
\code{\link{fmTime$isInfinite}},
\code{\link{fmTime$isSinceEver}},
\code{\link{fmTime$round}},
\code{\link{fmTime$sub}},
\code{\link{fmTimeCreate}()},
\code{\link{fmTimeForever}()},
\code{\link{fmTimeParse}()},
\code{\link{fmTimeSinceEver}()},
\code{\link{fmTime}}
}
\concept{fmTime methods}
