% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fmTimeSinceEver}
\alias{fmTimeSinceEver}
\title{The -∞ time}
\usage{
fmTimeSinceEver()
}
\value{
a \code{\link{fmTime}} representing -∞.
}
\description{
The -∞ time
}
\seealso{
Other fmTime methods: 
\code{\link{'<.Rcpp_fmTime'}},
\code{\link{'<=.Rcpp_fmTime'}},
\code{\link{'==.Rcpp_fmTime'}},
\code{\link{'>.Rcpp_fmTime'}},
\code{\link{'>=.Rcpp_fmTime'}},
\code{\link{fmTime$add}},
\code{\link{fmTime$after}},
\code{\link{fmTime$before}},
\code{\link{fmTime$equals}},
\code{\link{fmTime$round}},
\code{\link{fmTime$sub}},
\code{\link{fmTimeCreate}()},
\code{\link{fmTimeForever}()},
\code{\link{fmTimeNow}()},
\code{\link{fmTimeParse}()},
\code{\link{fmTime}}
}
\concept{fmTime methods}
