% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experiment.R
\name{fmTrackingDataDirectoryInfo}
\alias{fmTrackingDataDirectoryInfo}
\title{a reference class that summarizes information about an
    output directory of FORT}
\description{
wrapper for
    \href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::TrackingDataDirectoryInfo/fort::myrmidon::TrackingDataDirectoryInfo}{fort::myrmidon::TrackingDataDirectoryInfo}
}
\section{Fields}{

\describe{
\item{\code{uri}}{the URI used in FORT studio}

\item{\code{absoluteFilePath}}{actual location of the directory on the filesystem}

\item{\code{frames}}{the number of stored frame in the directory}

\item{\code{start}}{the first captured \code{\link{fmTime}} in the directory}

\item{\code{end}}{the last captured \code{\link{fmTime}} in the directory}
}}

