% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen.R
\name{fmVector2dList$as.data.frame}
\alias{fmVector2dList$as.data.frame}
\title{Returns a copy of the list as a data.frame}
\value{
a copy of the list data organized in a data.frame
}
\description{
Returns a copy of the list as a data.frame

Returns a copy of the list as a data.frame. Also works with
the \code{as.data.frame} function
}
\examples{
    l <-fmVector2dListCreate(data.frame("x" = c(1,2,3), "y"= c(1,2,3)))
    as.data.frame(l)
    l$as.data.frame()
}
\seealso{
Other fmVector2dList methods: 
\code{\link{fmVector2dList$[[<-}},
\code{\link{fmVector2dList$[[}},
\code{\link{fmVector2dList$as.list}},
\code{\link{fmVector2dList$size}},
\code{\link{fmVector2dListCreate}()},
\code{\link{fmVector2dList}}
}
\concept{fmVector2dList methods}
