% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen.R
\name{fmVector2dList$as.list}
\alias{fmVector2dList$as.list}
\title{Returns a copy of the list as a plain R list}
\value{
a copy of the list data organized in a list
}
\description{
Returns a copy of the list as a plain R list

Returns a copy of the list as a plain R list. Also works with
the \code{as.list()} function
}
\examples{
    l <-fmVector2dListCreate(data.frame("x" = c(1,2,3), "y"= c(1,2,3)))
    \dontrun{as.list(l)}
    l$as.list()
}
\seealso{
Other fmVector2dList methods: 
\code{\link{fmVector2dList$[[<-}},
\code{\link{fmVector2dList$[[}},
\code{\link{fmVector2dList$as.data.frame}},
\code{\link{fmVector2dList$size}},
\code{\link{fmVector2dListCreate}()},
\code{\link{fmVector2dList}}
}
\concept{fmVector2dList methods}
