% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen.R
\name{fmVector2dList$size}
\alias{fmVector2dList$size}
\title{returns the size of the Vector2dList}
\value{
the number of points in the list
}
\description{
returns the size of the Vector2dList. Works also with
the \code{length} function
}
\examples{
    l <-fmVector2dListCreate(data.frame("x" = c(1,2,3), "y"= c(1,2,3)))
    l$size()
    \dontrun{length(l)}
}
\seealso{
Other fmVector2dList methods: 
\code{\link{fmVector2dList$[[<-}},
\code{\link{fmVector2dList$[[}},
\code{\link{fmVector2dList$as.data.frame}},
\code{\link{fmVector2dList$as.list}},
\code{\link{fmVector2dListCreate}()},
\code{\link{fmVector2dList}}
}
\concept{fmVector2dList methods}
