% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen.R
\name{fmVector2dList}
\alias{fmVector2dList}
\title{Represents a list of 2D point}
\description{
an object that is somewhat similar to a
list. fmVector2dList can take a list of numerical vector or a
dataframe with 'x', and 'y' column as an argument to create a
fmVector2dList. Function and methods expecting a fmVector2dList
will accept such list or data.frame as input.
}
\examples{
    l <-fmVector2dListCreate(data.frame("x" = c(1,2,3), "y"= c(1,2,3)))
    as.list(l) # will give a list of numerical vector
    as.data.frame(l) # will give a dataframe
}
\seealso{
Other fmVector2dList methods: 
\code{\link{fmVector2dList$[[<-}},
\code{\link{fmVector2dList$[[}},
\code{\link{fmVector2dList$as.data.frame}},
\code{\link{fmVector2dList$as.list}},
\code{\link{fmVector2dList$size}},
\code{\link{fmVector2dListCreate}()}
}
\concept{fmVector2dList methods}
