% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fmVector2dListCreate}
\alias{fmVector2dListCreate}
\title{Creates a fmVector2dList from a list of numerical vector or a data.frame}
\usage{
fmVector2dListCreate(points = NULL)
}
\arguments{
\item{points}{a list of numerical vector of size 2 or a data.frame with column 'x' and 'y'}
}
\value{
a fmVector2dList
}
\description{
Creates a fmVector2dList from a list of numerical vector or a data.frame
}
\seealso{
Other fmVector2dList methods: 
\code{\link{fmVector2dList$[[<-}},
\code{\link{fmVector2dList$[[}},
\code{\link{fmVector2dList$as.data.frame}},
\code{\link{fmVector2dList$as.list}},
\code{\link{fmVector2dList$size}},
\code{\link{fmVector2dList}}
}
\concept{fmVector2dList methods}
