% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone.R
\name{fmZone$addDefinition}
\alias{fmZone$addDefinition}
\title{Adds a new fmZoneDefinition to the fmZone.}
\arguments{
\item{shapes}{the geometry}

\item{start}{the first valid \link{\code{fmTime}} for this definition.}

\item{end}{the first invalid \link{\code{fmTime}} for this definition.}
}
\value{
a \link{\code{fmZoneDefinition}}
}
\description{
Adds a new \link{\code{fmZoneDefinition}} to the
fmZone.
}
\seealso{
Other fmZone classes and methods: 
\code{\link{fmZone$deleteDefinition}},
\code{\link{fmZoneDefinition}},
\code{\link{fmZone}}
}
\concept{fmZone classes and methods}
