% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone.R
\name{fmZone$deleteDefinition}
\alias{fmZone$deleteDefinition}
\title{Removes a fmZoneDefinition to the fmZone.}
\arguments{
\item{index}{the index of the definition to remove}
}
\description{
Removes a \link{\code{fmZoneDefinition}} to the
fmZone.
}
\seealso{
Other fmZone classes and methods: 
\code{\link{fmZone$addDefinition}},
\code{\link{fmZoneDefinition}},
\code{\link{fmZone}}
}
\concept{fmZone classes and methods}
