% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone.R
\name{fmZone}
\alias{fmZone}
\title{A geometric zone of a space}
\description{
A fmZone represents a geometric zone of a
\code{\link{fmSpace}} for tracking localization and
interactions. It means that two Ant, which are lying in two
separate Zone will never report a collision or Interaction.

fmZone can only be created from a
\code{\link{fmSpace}} with \code{\link{fmSpace$createZone}()}.
}
\details{
Zones are uniquely idenrified trough their ID in an
\code{\link{fmExperiment}}, but they are managed and are
related to a single \code{\link{fmSpace}}. They also have a
user defined name, but it is never used internally.

By default an \code{\link{fmAnt}} lies in no fmZone, which
is identified by the ZoneID 0.

Zone have time valid \code{\link{fmZoneDefinition}} which
represents their geometry. In most cases a Zone will have a
single ZoneDefinition valid for
]\code{\link{fmTimeSinceEver}()};\code{\link{fmTimeForever}()}[. However
it is possible to add as many different definition to a Zone,
as long as they do not overlap in Time. The defintions are
manipulated with \code{\link{fmZone$addDefinition}()} and
\code{\link{fmZone$deleteDefinition}()}.
}
\section{Fields}{

\describe{
\item{\code{name}}{the zone name}

\item{\code{ID}}{the zone ID, which is unique in the experiment}

\item{\code{definitions}}{(read-only) the zone definitions}
}}

\seealso{
Other fmZone classes and methods: 
\code{\link{fmZone$addDefinition}},
\code{\link{fmZone$deleteDefinition}},
\code{\link{fmZoneDefinition}}
}
\concept{fmZone classes and methods}
