\name{FortMyrmidon-package}
\alias{FortMyrmidon-package}
\alias{FortMyrmidon}
\docType{package}
\title{
\packageTitle{FortMyrmidon}
}
\description{
\packageDescription{FortMyrmidon}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{FortMyrmidon}
\packageIndices{FortMyrmidon}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{FortMyrmidon}

Maintainer: \packageMaintainer{FortMyrmidon}
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
