% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{FortMyrmidon}
\alias{FortMyrmidon}
\title{FortMyrmidon}
\description{
R bindings to \strong{fort-myrmidon}. Its is an object
oriented API, please look up the following classes bellow:

\itemize{
\item API entry point: \code{\link{fmExperiment}}
\item Ants and Identification: \itemize{ \item \code{\link{fmAnt}} \item \code{\link{fmIdentification}} }
\item Spaces and Zones: \itemize{ \item \code{\link{fmSpace}} \item \code{\link{fmZone}} \item \code{\link{fmZoneDefinition}} }
\item Queries and Marchers: \itemize{ \item \code{\link{fmQuery}} \item \code{\link{fmMatcher}} }
\item Time representation: \itemize{ \item \code{\link{fmTime}} \item \code{\link{fmDuration}} }
\item 2D Shapes : \itemize{ \item \code{\link{fmCircle}} \item \code{\link{fmCapsule}} \item \code{\link{fmPolygon}} }
}
}
\author{
alexandre.tuleu.2005@polytechnique.org
}
