% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duration.R
\name{>=.Rcpp_fmDuration}
\alias{>=.Rcpp_fmDuration}
\title{Comparison operator for fmDuration}
\usage{
\method{>=}{Rcpp_fmDuration}(e1, e2)
}
\arguments{
\item{e1}{the first fmDuration to compare}

\item{e2}{the second fmDuration to compare}
}
\value{
\code{TRUE} if \code{e1 >= e2}
}
\description{
Comparison operator for fmDuration
}
\examples{
\dontrun{
fmMinute(1) >= fmSecond(1)
# outputs TRUE
fmMinute(60) >= fmSecond(60)
# outputs TRUE (as 60s == 1m)
}
}
