% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duration.R
\name{!=.Rcpp_fmDuration}
\alias{!=.Rcpp_fmDuration}
\title{Different operator for fmDuration}
\usage{
\method{!=}{Rcpp_fmDuration}(e1, e2)
}
\arguments{
\item{e1}{the first fmDuration to compare}

\item{e2}{the second fmDuration to compare}
}
\value{
\code{TRUE} if \code{e1 != e2}
}
\description{
Different operator for fmDuration
}
\examples{
\dontrun{
fmSecond(1) != fmMinute(1)
# outputs TRUE
fmSecond(60) != fmMinute(1)
# outputs FALSE (as 60s == 1m)
}
}
