% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time.R
\name{!=.Rcpp_fmTime}
\alias{!=.Rcpp_fmTime}
\title{Difference operator for fmTime}
\usage{
\method{!=}{Rcpp_fmTime}(e1, e2)
}
\arguments{
\item{e1}{the first fmTime to compare}

\item{e2}{the second fmTime to compare}
}
\value{
\code{TRUE} if \code{e1} and  \code{e2} do ont represents the same time at nanoseconds precision
}
\description{
Difference operator for fmTime
}
\examples{
\dontrun{
fmTimeParse("2020-01-29T08:00:00.000Z") == fmTimeParse("2020-01-29T08:00:00.000Z")
# outputs FALSE
fmTimeParse("2020-01-29T08:00:00.000Z") == fmTimeParse("2020-01-29T08:00:00.000Z")$add(1)
# outpus TRUE (1ns difference)
}
}
