% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duration.R
\name{+.Rcpp_fmDuration}
\alias{+.Rcpp_fmDuration}
\title{Addition operator for fmDuration}
\usage{
\method{+}{Rcpp_fmDuration}(e1, e2)
}
\arguments{
\item{e1}{the first fmDuration to add}

\item{e2}{the second fmDuration to add}
}
\value{
the sum of the two fmDuration
}
\description{
Addition operator for fmDuration
}
\examples{
\dontrun{
fmSecond(2) + fmMicrosecond(300)
# outputs 2.0003s
}
}
