% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{Rcpp_fmTime-class}
\alias{Rcpp_fmTime-class}
\title{Internal Object}
\description{
representation for \code{\link{fmTime}}
}
\seealso{
Other internals: 
\code{\link{Rcpp_fmDuration-class}}
}
\concept{internals}
