% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{'<=.Rcpp_fmDuration'}
\alias{'<=.Rcpp_fmDuration'}
\title{Comparison operator for fmDuration}
\arguments{
\item{e1}{the first fmDuration to compare}

\item{e2}{the second fmDuration to compare}
}
\value{
\code{TRUE} if \code{e1 <= e2}
}
\description{
Comparison operator for fmDuration
}
\examples{
\dontrun{
fmSecond(1) <= fmMinute(1)
fmSecond(60) <= fmMinute(1)
}
}
\seealso{
Other fmDuration methods: 
\code{\link{'+.Rcpp_fmDuration'}},
\code{\link{'-.Rcpp_fmDuration'}},
\code{\link{'<.Rcpp_fmDuration'}},
\code{\link{'==.Rcpp_fmDuration'}},
\code{\link{'>.Rcpp_fmDuration'}},
\code{\link{'>=.Rcpp_fmDuration'}},
\code{\link{fmDurationParse}()},
\code{\link{fmDuration}},
\code{\link{fmHour}()},
\code{\link{fmMicrosecond}()},
\code{\link{fmMillisecond}()},
\code{\link{fmMinute}()},
\code{\link{fmNanosecond}()},
\code{\link{fmSecond}()}
}
\concept{fmDuration methods}
