% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{'+.Rcpp_fmDuration'}
\alias{'+.Rcpp_fmDuration'}
\title{Addition operator for fmDuration}
\arguments{
\item{e1}{the first fmDuration to add}

\item{e2}{the second fmDuration to add}
}
\value{
the sum of the two fmDuration
}
\description{
Addition operator for fmDuration
}
\examples{
\dontrun{
fmSecond(2) + fmMicrosecond(300)
}
}
\seealso{
Other fmDuration methods: 
\code{\link{'-.Rcpp_fmDuration'}},
\code{\link{'<.Rcpp_fmDuration'}},
\code{\link{'<=.Rcpp_fmDuration'}},
\code{\link{'==.Rcpp_fmDuration'}},
\code{\link{'>.Rcpp_fmDuration'}},
\code{\link{'>=.Rcpp_fmDuration'}},
\code{\link{fmDurationParse}()},
\code{\link{fmDuration}},
\code{\link{fmHour}()},
\code{\link{fmMicrosecond}()},
\code{\link{fmMillisecond}()},
\code{\link{fmMinute}()},
\code{\link{fmNanosecond}()},
\code{\link{fmSecond}()}
}
\concept{fmDuration methods}
