% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ant.R
\name{fmAnt}
\alias{fmAnt}
\title{A reference class that represent an Ant}
\description{
This is a wrapper for the C++ class
    \href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Ant}{fort::myrmidon::Ant}
}
\section{Fields}{

\describe{
\item{\code{cIdentifications}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Ant/fort::myrmidon::Ant::CIdentifications}{fort::myrmidon::Ant::CIdentifications}}

\item{\code{identifications}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Ant/fort::myrmidon::Ant::Identifications}{fort::myrmidon::Ant::Identifications}}

\item{\code{antID}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Ant/fort::myrmidon::Ant::AntID}{fort::myrmidon::Ant::AntID}}

\item{\code{formattedID}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Ant/fort::myrmidon::Ant::FormattedID}{fort::myrmidon::Ant::FormattedID}}

\item{\code{displayColor}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Ant/fort::myrmidon::Ant::DisplayColor}{fort::myrmidon::Ant::DisplayColor}}

\item{\code{setDisplayColor}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Ant/fort::myrmidon::Ant::SetDisplayColor}{fort::myrmidon::Ant::SetDisplayColor}}

\item{\code{displayStatus}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Ant/fort::myrmidon::Ant::DisplayStatus}{fort::myrmidon::Ant::DisplayStatus}}

\item{\code{getValue}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Ant/fort::myrmidon::Ant::GetValue}{fort::myrmidon::Ant::GetValue}}

\item{\code{deleteValue}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Ant/fort::myrmidon::Ant::DeleteValue}{fort::myrmidon::Ant::DeleteValue}}

\item{\code{setValue}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Ant/fort::myrmidon::Ant::SetValue}{fort::myrmidon::Ant::SetValue}}

\item{\code{addCapsule}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Ant/fort::myrmidon::Ant::AddCapsule}{fort::myrmidon::Ant::AddCapsule}}

\item{\code{capsules}}{wrapper for
\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Ant/fort::myrmidon::Ant::Capsules}{fort::myrmidon::Ant::Capsules}. It
returns a list with two elements, the 'capsules' element, a
list of \linkS4class{fmCapsule} and 'shapeTypes' their
corresponding shape types.}

\item{\code{deleteCapsule}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Ant/fort::myrmidon::Ant::DeleteCapsule}{fort::myrmidon::Ant::DeleteCapsule}}

\item{\code{clearCapsules}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Ant/fort::myrmidon::Ant::ClearCapsules}{fort::myrmidon::Ant::ClearCapsules}}

\item{\code{setDisplayStatus}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Ant/fort::myrmidon::Ant::SetDisplayStatus}{fort::myrmidon::Ant::SetDisplayStatus}}
}}

