% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shapes.R
\name{fmCapsule}
\alias{fmCapsule}
\title{a Reference Class that wraps a C++ \href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Capsule/fort::myrmidon::Capsule}{fort::myrmidon::Capsule}}
\description{
a Reference Class that wraps a C++ \href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Capsule/fort::myrmidon::Capsule}{fort::myrmidon::Capsule}
}
\section{Fields}{

\describe{
\item{\code{c1}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Capsule/fort::myrmidon::Capsule::C1}{fort::myrmidon::Capsule::C1}}

\item{\code{c2}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Capsule/fort::myrmidon::Capsule::C2}{fort::myrmidon::Capsule::C2}}

\item{\code{r1}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Capsule/fort::myrmidon::Capsule::R1}{fort::myrmidon::Capsule::R1}}

\item{\code{r2}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Capsule/fort::myrmidon::Capsule::R2}{fort::myrmidon::Capsule::R2}}

\item{\code{setC1}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Capsule/fort::myrmidon::Capsule::SetC1}{fort::myrmidon::Capsule::SetC1}}

\item{\code{setC2}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Capsule/fort::myrmidon::Capsule::SetC2}{fort::myrmidon::Capsule::SetC2}}

\item{\code{setR1}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Capsule/fort::myrmidon::Capsule::SetR1}{fort::myrmidon::Capsule::SetR1}}

\item{\code{setR2}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Capsule/fort::myrmidon::Capsule::SetR2}{fort::myrmidon::Capsule::SetR2}}
}}

\examples{
# creates a fmCapsule of center (1,1) and (2,2) radius 0.5 and 0.4
c <- fmCapsuleCreate(c(1,1),c(2,2),0.5,0.4)
# moves second center to (0,0)
c$setC2(c(0,0))
}
