% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shapes.R
\name{fmCircle}
\alias{fmCircle}
\title{a Reference Class that wraps a C++ \href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Circle/fort::myrmidon::Circle}{fort::myrmidon::Circle}}
\description{
a Reference Class that wraps a C++ \href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Circle/fort::myrmidon::Circle}{fort::myrmidon::Circle}
}
\section{Fields}{

\describe{
\item{\code{center}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Circle/fort::myrmidon::Circle::Center}{fort::myrmidon::Circle::Center}}

\item{\code{radius}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Circle/fort::myrmidon::Circle::Radius}{fort::myrmidon::Circle::Radius}}

\item{\code{setCenter}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Circle/fort::myrmidon::Circle::SetCenter}{fort::myrmidon::Circle::SetSetRadius}}

\item{\code{setRadius}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Circle/fort::myrmidon::Circle::SetRadius}{fort::myrmidon::Circle::SetRadius}}
}}

\examples{
# creates a fmCircle of center 1,1 amnd radius 2
c <- fmCircleCreate(c(1,1),2)
# moves it to (0,0)
c$setCenter(c(0,0))

}
