% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{fmCircleCreate}
\alias{fmCircleCreate}
\title{Creates a circle}
\usage{
fmCircleCreate(center = c(0, 0), radius = 1)
}
\arguments{
\item{center}{the circle's center}

\item{radius}{the circle's radius}
}
\value{
a \code{\link{fmCircle}}
}
\description{
Creates a circle
}
\seealso{
Other fmShape methods and classes: 
\code{\link{fmCapsuleCreate}()},
\code{\link{fmCapsule}},
\code{\link{fmCircle}},
\code{\link{fmPolygonCreate}()},
\code{\link{fmPolygon}},
\code{\link{fmShapeList$[[<-}},
\code{\link{fmShapeList$[[}},
\code{\link{fmShapeList$as.list}},
\code{\link{fmShapeList$size}},
\code{\link{fmShapeListCreate}()},
\code{\link{fmShapeList}},
\code{\link{fmShapeTypes}},
\code{\link{fmShape}}
}
\concept{fmShape methods and classes}
