% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Color.R
\name{fmDefaultPaletteColor}
\alias{fmDefaultPaletteColor}
\title{Creates an \code{\link{fmColor}} from myrmidon palette}
\arguments{
\item{i}{an integer to draw a color from}
}
\value{
the wanted \code{\link{fmColor}}
}
\description{
myrmidon defines a color-blind friendly palette of 7
    colors that can be accessed with this method. #i does not need
    to be between 0 and 6 color will cycle through the 7 available
    colors.
}
\examples{
# the two fmColor below will be distinguishable for most color-blind people
firstColor <- fmDefaultPaletteColor(0)
secondColor <- fmDefaultPaletteColor(1)
}
