% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duration.R
\name{fmDuration}
\alias{fmDuration}
\title{A Reference Class that represent a duration}
\description{
This a wrapper for \href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Duration/fort::myrmidon::Duration}{fort::myrmidon::Duration}
}
\section{Fields}{

\describe{
\item{\code{hours}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Duration/fort::myrmidon::Duration::Hours}{fort::myrmidon::Duration::Hours}}

\item{\code{minutes}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Duration/fort::myrmidon::Duration::Minutes}{fort::myrmidon::Duration::Minutes}}

\item{\code{seconds}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Duration/fort::myrmidon::Duration::Seconds}{fort::myrmidon::Duration::Seconds}}

\item{\code{milliseconds}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Duration/fort::myrmidon::Duration::Milliseconds}{fort::myrmidon::Duration::Milliseconds}}

\item{\code{microseconds}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Duration/fort::myrmidon::Duration::Microseconds}{fort::myrmidon::Duration::Microseconds}}

\item{\code{Nanoseconds}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Duration/fort::myrmidon::Duration::Nanoseconds}{fort::myrmidon::Duration::Nanoseconds}}
}}

