% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Experiment.R
\name{fmExperiment}
\alias{fmExperiment}
\title{a reference class wrapper for the C++
    \href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment}{fort::myrmidon::Experiment}
    class.}
\description{
a reference class wrapper for the C++
    \href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment}{fort::myrmidon::Experiment}
    class.
}
\section{Fields}{

\describe{
\item{\code{const}}{return a const version, i.e. \code{\link{fmCExperiment}} for this fmExperiment}

\item{\code{save}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::Save}{fort::myrmidon::Experiment::Save}}

\item{\code{absoluteFilePath}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::AbsoluteFilePath}{fort::myrmidon::Experiment::AbsoluteFilePath}}

\item{\code{createSpace}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::CreateSpace}{fort::myrmidon::Experiment::CreateSpace}}

\item{\code{deleteSpace}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::DeleteSpace}{fort::myrmidon::Experiment::DeleteSpace}}

\item{\code{cSpaces}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::CSpaces}{fort::myrmidon::Experiment::CSpaces}}

\item{\code{spaces}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::Spaces}{fort::myrmidon::Experiment::Spaces}}

\item{\code{addTrackingDataDirectory}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::AddTrackingDataDirectory}{fort::myrmidon::Experiment::AddTrackingDataDirectory}}

\item{\code{deleteTrackingDataDirectory}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::DeleteTrackingDataDirectory}{fort::myrmidon::Experiment::DeleteTrackingDataDirectory}}

\item{\code{createAnt}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::CreateAnt}{fort::myrmidon::Experiment::CreateAnt}}

\item{\code{cAnts}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::CAnts}{fort::myrmidon::Experiment::CAnts}}

\item{\code{ants}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::Ants}{fort::myrmidon::Experiment::Ants}}

\item{\code{addIdentification}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::AddIdentification}{fort::myrmidon::Experiment::AddIdentification}}

\item{\code{deleteIdentification}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::DeleteIdentification}{fort::myrmidon::Experiment::DeleteIdentification}}

\item{\code{freeIdentificationRangeAt}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::FreeIdentificationRangeAt}{fort::myrmidon::Experiment::FreeIdentificationRangeAt}}

\item{\code{name}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::Name}{fort::myrmidon::Experiment::Name}}

\item{\code{setName}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::SetName}{fort::myrmidon::Experiment::SetName}}

\item{\code{comment}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::Comment}{fort::myrmidon::Experiment::Comment}}

\item{\code{setComment}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::SetComment}{fort::myrmidon::Experiment::SetComment}}

\item{\code{author}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::Author}{fort::myrmidon::Experiment::Author}}

\item{\code{setAuthor}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::SetAuthor}{fort::myrmidon::Experiment::SetAuthor}}

\item{\code{family}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::Family}{fort::myrmidon::Experiment::Family}}

\item{\code{setFamily}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::SetFamily}{fort::myrmidon::Experiment::SetFamily}}

\item{\code{defaultTagSize}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::DefaultTagSize}{fort::myrmidon::Experiment::DefaultTagSize}}

\item{\code{setDefaultTagSize}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::SetDefaultTagSize}{fort::myrmidon::Experiment::SetDefaultTagSize}}

\item{\code{threshold}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::Threshold}{fort::myrmidon::Experiment::Threshold}}

\item{\code{setThreshold}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::SetThreshold}{fort::myrmidon::Experiment::SetThreshold}}

\item{\code{createMeasurementType}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::CreateMeasurementType}{fort::myrmidon::Experiment::CreateMeasurementType}}

\item{\code{measurementTypeNames}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::MeasurementTypeNames}{fort::myrmidon::Experiment::MeasurementTypeNames}}

\item{\code{deleteMeasurementType}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::DeleteMeasurementType}{fort::myrmidon::Experiment::DeleteMeasurementType}}

\item{\code{setMeasurementTypeName}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::SetMeasurementTypeName}{fort::myrmidon::Experiment::SetMeasurementTypeName}}

\item{\code{createAntShapeType}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::CreateAntShapeType}{fort::myrmidon::Experiment::CreateAntShapeType}}

\item{\code{antShapeTypeNames}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::AntShapeTypeNames}{fort::myrmidon::Experiment::AntShapeTypeNames}}

\item{\code{setAntShapeTypeName}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::SetAntShapeTypeName}{fort::myrmidon::Experiment::SetAntShapeTypeName}}

\item{\code{deleteAntShapeType}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::DeleteAntShapeType}{fort::myrmidon::Experiment::DeleteAntShapeType}}

\item{\code{addMetadataColumn}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::AddMetadataColumn}{fort::myrmidon::Experiment::AddMetadataColumn}}

\item{\code{deleteMetadataColumn}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::DeleteMetadataColumn}{fort::myrmidon::Experiment::DeleteMetadataColumn}}

\item{\code{antMetadataColumns}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::AntMetadataColumns}{fort::myrmidon::Experiment::AntMetadataColumns}}

\item{\code{renameAntMetadataColumn}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::RenameAntMetadataColumn}{fort::myrmidon::Experiment::RenameAntMetadataColumn}}

\item{\code{setAntMetadataColumnType}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::SetAntMetadataColumnType}{fort::myrmidon::Experiment::SetAntMetadataColumnType}}

\item{\code{getDataInformations}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Experiment/fort::myrmidon::Experiment::GetDataInformations}{fort::myrmidon::Experiment::GetDataInformations}}
}}

