% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fmExperimentCreate}
\alias{fmExperimentCreate}
\title{Creates an Experiement}
\usage{
fmExperimentCreate(filepath)
}
\arguments{
\item{filepath}{the name for the newly created space}
}
\value{
a \link{\code{fmExperiment}} with name
}
\description{
Creates an Experiement
}
