% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duration.R
\name{fmHour}
\alias{fmHour}
\title{Creates a fmDuration which is a multiple of an hour}
\arguments{
\item{v}{the wanted number of hours}
}
\value{
a fmDuration representing the wanted amount of hours
}
\description{
Creates a fmDuration which is a multiple of an hour
}
\examples{
fmHour(2)
# outputs 2h
fmHour(1.5)
# outputs 1h30m
}
