% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identification.R
\name{fmIdentification}
\alias{fmIdentification}
\title{Reference Class wrapper for C++ \href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Identification/fort::myrmidon::Identification}{fort::myrmidon::Identification}}
\description{
Reference Class wrapper for C++ \href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Identification/fort::myrmidon::Identification}{fort::myrmidon::Identification}
}
\section{Fields}{

\describe{
\item{\code{tagValue}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Identification/fort::myrmidon::Identification::TagValue}{fort::myrmidon::Identification::TagValue}}

\item{\code{targetAntID}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Identification/fort::myrmidon::Identification::TargetAntID}{fort::myrmidon::Identification::TargetAntID}}

\item{\code{setStart}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Identification/fort::myrmidon::Identification::SetStart}{fort::myrmidon::Identification::SetStart}}

\item{\code{setEnd}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Identification/fort::myrmidon::Identification::SetEnd}{fort::myrmidon::Identification::SetEnd}}

\item{\code{start}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Identification/fort::myrmidon::Identification::Start}{fort::myrmidon::Identification::Start}}

\item{\code{end}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Identification/fort::myrmidon::Identification::End}{fort::myrmidon::Identification::End}}

\item{\code{antPosition}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Identification/fort::myrmidon::Identification::AntPosition}{fort::myrmidon::Identification::AntPosition}}

\item{\code{antAngle}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Identification/fort::myrmidon::Identification::AntAngle}{fort::myrmidon::Identification::AntAngle}}

\item{\code{hasUserDefinedAntPose}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Identification/fort::myrmidon::Identification::HasUserDefinedAntPose}{fort::myrmidon::Identification::HasUserDefinedAntPose}}

\item{\code{setUserDefinedAntPosition}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Identification/fort::myrmidon::Identification::SetUserDefinedAntPosition}{fort::myrmidon::Identification::SetUserDefinedAntPosition}}

\item{\code{clearUserDefinedAntPose}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Identification/fort::myrmidon::Identification::ClearUserDefinedAntPose}{fort::myrmidon::Identification::ClearUserDefinedAntPose}}
}}

