% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matchers.R
\name{fmMatcher}
\alias{fmMatcher}
\title{Reference abstract class to match criterion on queries}
\description{
fmMatchers can be used to match against certain
    criterion, here are the available matcher:
\itemize{
\item \code{\link{fmMatcherAny}}
\item \code{\link{fmMatcherAnd}}
\item \code{\link{fmMatcherOr}}
\item \code{\link{fmMatcherAntID}}
\item \code{\link{fmMatcherAntColumn}}
\item \code{\link{fmMatcherAntDistanceGreaterThan}}
\item \code{\link{fmMatcherAntDistanceSmallerThan}}
\item \code{\link{fmMatcherAntAngleSmallerThan}}
\item \code{\link{fmMatcherAntAngleGreaterThan}}
}
}
