% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matchers.R
\name{fmMatcherAnd}
\alias{fmMatcherAnd}
\title{\code{\link{fmMatcher}} a matcher that matches all other
    matchers}
\arguments{
\item{matchers}{a list of other \code{\link{fmMatcher}} to match}
}
\value{
a \code{\link{fmMatcher}} a matcher that matches when all
    subMatcher matches
}
\description{
\code{\link{fmMatcher}} a matcher that matches all other
    matchers
}
\examples{
# a fmMatcher that matches two criterion (could be more)
fmMatcherAnd(list(fmMatcherAntColumn("group","worker"),fmMatcherAntID(42)))
}
