% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matchers.R
\name{fmMatcherAntID}
\alias{fmMatcherAntID}
\title{\code{\link{fmMatcher}} a matcher that matches ant with
    given ID}
\arguments{
\item{ID}{the ID to match}
}
\value{
a \code{\link{fmMatcher}} a matcher that matches when any
    ant has the given ID
}
\description{
\code{\link{fmMatcher}} a matcher that matches ant with
    given ID
}
\examples{
# a fmMatcher that matches ant ID 42
fmMatcherAntID(42)
}
