% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matchers.R
\name{fmMatcherAny}
\alias{fmMatcherAny}
\title{\code{\link{fmMatcherAny}} a matcher that matches anything}
\value{
a \code{\link{fmMatcher}} a matcher that matches anything
}
\description{
\code{\link{fmMatcherAny}} a matcher that matches anything
}
