% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matchers.R
\name{fmMatcherInteractionType}
\alias{fmMatcherInteractionType}
\title{\code{\link{fmMatcher}} a matcher that matches when an
    interaction contains the given type}
\arguments{
\item{type1}{the first virtual ant shape type to match}

\item{type2}{the second virtual ant shape type to match}
}
\value{
a \code{\link{fmMatcher}} a matcher that matches when the
    interaction contains the given type. Matches (type1,type2) and
    (type2,type1). Only valid for query on collision or
    interaction, and always matches for position or trajectory
    queries.
}
\description{
\code{\link{fmMatcher}} a matcher that matches when an
    interaction contains the given type
}
\examples{
# matches 1 <-> 2 and 2 <-> 1 interactions
fmMatcherInteractionType(1,2)
}
