% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matchers.R
\name{fmMatcherOr}
\alias{fmMatcherOr}
\title{\code{\link{fmMatcher}} a matcher that matches any other
    matchers}
\arguments{
\item{matchers}{a list of other \code{\link{fmMatcher}} to match}
}
\value{
a \code{\link{fmMatcher}} a matcher that matches when any
    sub-fmMatcher matches
}
\description{
\code{\link{fmMatcher}} a matcher that matches any other
    matchers
}
\examples{
# a fmMatcher that matches any two criterion (could be more)
fmMatcherOr(list(fmMatcherAntColumn("group","workers"),fmMatcherAntColumn("group","nurses")))
}
