% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duration.R
\name{fmMicrosecond}
\alias{fmMicrosecond}
\title{Creates a fmDuration which is a multiple of a microseconds}
\arguments{
\item{v}{the wanted number of microseconds}
}
\value{
a fmDuration representing the wanted amount of microseconds
}
\description{
Creates a fmDuration which is a multiple of a microseconds
}
\examples{
fmMicrosecond(2)
# outputs 2us
fmMinute(1.5)
# outputs 1.5us
}
