% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duration.R
\name{fmMillisecond}
\alias{fmMillisecond}
\title{Creates a fmDuration which is a multiple of a millisecond}
\arguments{
\item{v}{the wanted number of milliseconds}
}
\value{
a fmDuration representing the wanted amount of milliseconds
}
\description{
Creates a fmDuration which is a multiple of a millisecond
}
\examples{
fmMillisecond(2)
# outputs 2ms
fmMillisecond(1.5)
# outputs 1.5ms
}
