% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duration.R
\name{fmMinute}
\alias{fmMinute}
\title{Creates a fmDuration which is a multiple of a minute}
\arguments{
\item{v}{the wanted number of minutes}
}
\value{
a fmDuration representing the wanted amount of minutes
}
\description{
Creates a fmDuration which is a multiple of a minute
}
\examples{
fmMinute(2)
# outputs 2m
fmMinute(1.5)
# outputs 1m30s
}
