% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shapes.R
\name{fmPolygon}
\alias{fmPolygon}
\title{a Reference Class that wraps a C++ \href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Polygon/fort::myrmidon::Capsule}{fort::myrmidon::Polygon}}
\description{
a Reference Class that wraps a C++ \href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Polygon/fort::myrmidon::Capsule}{fort::myrmidon::Polygon}
}
\section{Fields}{

\describe{
\item{\code{size}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Polygon/fort::myrmidon::Polygon::Size}{fort::myrmidon::Polygon::Size}}

\item{\code{vertex}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Polygon/fort::myrmidon::Polygon::Vertex}{fort::myrmidon::Polygon::Vertex}}

\item{\code{setVertex}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Polygon/fort::myrmidon::Polygon::SetVertex}{fort::myrmidon::Polygon::SetVertex}}
}}

\examples{
#creates a nice (0,0) centered square of size 2
p <- fmPolygonCreate(list(c(-1,-1),c(1,-1),c(1,1),c(-1,1)))
# transforms it into the extinction rebellion symbol by swapping the
# two lastelements, as order matters here.
# ATTENTION  list index starts at 0, like in C
p$setVertex(2,c(-1,1))
p$setVertex(3,c(1,1))
}
