% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{fmPolygonCreate}
\alias{fmPolygonCreate}
\title{Creates a polygon}
\usage{
fmPolygonCreate(vertices = NULL)
}
\arguments{
\item{vertices}{the polygon vertices}
}
\value{
a \code{\link{fmPolygon}}
}
\description{
Creates a polygon
}
\seealso{
Other fmShape methods and classes: 
\code{\link{fmCapsuleCreate}()},
\code{\link{fmCapsule}},
\code{\link{fmCircleCreate}()},
\code{\link{fmCircle}},
\code{\link{fmPolygon}},
\code{\link{fmShapeList$[[<-}},
\code{\link{fmShapeList$[[}},
\code{\link{fmShapeList$as.list}},
\code{\link{fmShapeList$size}},
\code{\link{fmShapeListCreate}()},
\code{\link{fmShapeList}},
\code{\link{fmShapeTypes}},
\code{\link{fmShape}}
}
\concept{fmShape methods and classes}
