% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Queries.R
\name{fmQueryCollideFrames}
\alias{fmQueryCollideFrames}
\title{Computes ant position and collision during the experiment}
\usage{
fmQueryCollideFrames(
  experiment,
  start = NULL,
  end = NULL,
  singleThreaded = FALSE,
  showProgress = FALSE
)
}
\arguments{
\item{experiment}{the \code{\link{fmCExperiment}} to query for, use
\code{e$const} if you didn't opened the experiment in read
only mode.}

\item{start}{the starting time for the query. Data which were
acquire before that time will not be reported. Could be NULL
for -∞, a POSIXct a fmTime or fmTimeCPtr object.}

\item{end}{the starting time for the query. Data which were acquire
after that time will not be reported. Could be NULL for +∞, a
POSIXct a fmTime or fmTimeCPtr object/}

\item{singleThreaded}{perform a single threaded query. Could be
faster for small experiment in number of Ant as identification
is mostly IO bounded and the overhead induced by
multi-threading synchronization impacts performances.}

\item{showProgress}{display the progress of the computation on the
standard error output. It may not be portable behavior on all
OS.}
}
\value{
a list of two lists. \code{$positions} a list of
    \linkS4class{fmIdentifiedFrame}, and \code{$collisions} a list
    of \linkS4class{fmCollisionFrame}. Both list refers to the same
    time and are ordered in time.
}
\description{
Computes ant position and collision during the
    experiment. unlike \code{\link{fmQueryIdentifyFrames}} zone
    reporting will always occurs (it is precomputing steps for
    collision detection).
}
