% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Queries.R
\name{fmQueryComputeAntInteractions}
\alias{fmQueryComputeAntInteractions}
\title{Computes ant interactions during the experiment}
\usage{
fmQueryComputeAntInteractions(
  experiment,
  start = NULL,
  end = NULL,
  maximumGap = fmSecond(1),
  matcher = NULL,
  singleThreaded = FALSE,
  showProgress = FALSE,
  reportTrajectories = FALSE
)
}
\arguments{
\item{experiment}{the \code{\link{fmCExperiment}} to query for, use
\code{e$const} if you didn't opened the experiment in read only
mode.}

\item{start}{the starting time for the query. Data which were
acquire before that time will not be reported. Could be NULL
for -∞, a POSIXct a fmTime or fmTimeCPtr object.}

\item{end}{the ending time for the query. Data which were acquire
after or at that exact time will not be reported. Could be NULL
for +∞, a POSIXct a fmTime or fmTimeCPtr object}

\item{maximumGap}{the maximum gap in tracking before cutting the
trajectory in two different object. Use an insanely large value
( such as \code{fmHour(24*365)} to disable cutting of
trajectory). Trajectories and Interactions will always be cut
when an ant moves from a space to another.}

\item{matcher}{a \link{fmMatcher} to reduce the query to
wanted criterion. You can use \code{NULL} or
\code{\link{fmMatcherAny}()} to match anything. Matcher that
requires two ant will match conditionally only for the
interaction detection, but will always match for interaction
detection. Matcher that addresses a single ant will match
interaction where any of the two involved ant matches.}

\item{singleThreaded}{perform a single threaded query. Could be
faster for small experiment in number of Ant as identification
is mostly IO bounded and the overhead induced by
multi-threading synchronization impacts performances.}

\item{showProgress}{display the progress of the computation on the
standard error output. It may not be portable behavior on all
OS.}

\item{reportTrajectories}{enables global trajectory report in the
result. If enabled two lines in the resulting list will be
added, as described by the output of
\code{\link{fmQueryComputeAntTrajectories}}. Each interaction
will also refers to these two lines, providing the index of the
corresponding for each ant,}
}
\value{
a list with at list an element called \code{interactions}
    \code{data.frame} summarising all interactions. Other elements
    depends on the choice of othere report options.
}
\description{
Computes ant interactions and trajectories during the
    experiment. unlike \code{\link{fmQueryComputeAntTrajectory}} zone
    reporting will always occurs (it is a precomputing step for
    collision detection and therefore interaction computation).
}
