% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Queries.R
\name{fmQueryComputeAntTrajectories}
\alias{fmQueryComputeAntTrajectories}
\title{Computes ant trajectories during the experiment}
\usage{
fmQueryComputeAntTrajectories(
  experiment,
  start = NULL,
  end = NULL,
  maximumGap = fmSecond(1),
  matcher = NULL,
  computeZones = FALSE,
  singleThreaded = FALSE,
  showProgress = FALSE
)
}
\arguments{
\item{experiment}{the \code{\link{fmCExperiment}} to query for, use
\code{e$const} if you didn't opened the experiment in read only
mode.}

\item{start}{the starting time for the query. Data which were
acquire before that time will not be reported. Could be NULL
for -∞, a POSIXct a fmTime or fmTimeCPtr object.}

\item{end}{the ending time for the query. Data which were acquire
after or at that exact time will not be reported. Could be NULL
for +∞, a POSIXct a fmTime or fmTimeCPtr object}

\item{maximumGap}{the maximum gap in tracking before cutting the
trajectory in two different object. Use an insanely large value
( such as \code{fmHour(24*365)} to disable cutting of
trajectory). Trajectory will always be cut when an ant moves
from a space to another.}

\item{matcher}{a \link{fmMatcher} to reduce the query to
wanted criterion. You can use \code{NULL} or
\code{\link{fmMatcherAny}()} to match anything. Matcher that
requires two ant will match anything.}

\item{computeZones}{reports the user defined zone for the
ants. otherwise the data.frame will be smaller and the
computation faster.}

\item{singleThreaded}{perform a single threaded query. Could be
faster for small experiment in number of Ant as identification
is mostly IO bounded and the overhead induced by
multi-threading synchronization impacts performances.}

\item{showProgress}{display the progress of the computation on the
standard error output. It may not be portable behavior on all
OS.}
}
\value{
a list of two elements: a) \code{summaryTrajectories} a
    \code{data.frame} summarizing the ant, start time, and space of
    each trajectories reported, and b) \code{trajectories} a list
    of \code{data.frame} with the data point for every
    trajectories. Each element of this list correspond to the same
    row in the summary \code{data.frame}
}
\description{
Computes ant trajectories during the experiment
}
