% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Queries.R
\name{fmQueryComputeMeasurementFor}
\alias{fmQueryComputeMeasurementFor}
\title{Computes manual measurement for an Ant}
\arguments{
\item{experiment}{the \code{\link{fmCExperiment}} to query for, use
\code{e$const} if you didn't opened the experiment in read only
mode.}

\item{antID}{the ant we are interested in}

\item{mTypeID}{the measurement type id we are interceded in}
}
\value{
a data.frame with the date of the measurement and the
    length in mm.
}
\description{
access and computes the manual measurement made in the
    FORT Studio GUI. Length are reported in mm from the detected
    tag size and the tag size parameter. See
    \url{https://formicidae-tracker.github.io/studio/docs/latest/api/#manual_measurement/manual_measurement}
}
