% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fmQueryComputeTagStatistics}
\alias{fmQueryComputeTagStatistics}
\title{Computes tag statistics for an experiment}
\usage{
fmQueryComputeTagStatistics(experiment)
}
\arguments{
\item{experiment}{the \code{\link{fmExperiment}} to query}
}
\value{
a \code{data.frame} with the detection statistics in
\strong{experiment}
}
\description{
Computes tag statistics for an experiment
}
\seealso{
Other fmQuery methods: 
\code{\link{fmQueryComputeMeasurementFor}()},
\code{\link{fmQuery}}
}
\concept{fmQuery methods}
