% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Queries.R
\name{fmQueryComputeTagStatistics}
\alias{fmQueryComputeTagStatistics}
\title{Computes tag tracking statistics for an experiment}
\arguments{
\item{experiment}{the \code{\link{fmCExperiment}} to query for, use
\code{e$const} if you didn't opened the experiment in read only
mode.}
}
\value{
a data.frame with the statistics for each tag.
}
\description{
Computes tracking statistics by tag for an
    experiment. It consists of:
    \itemize{
    \item the dates at which a tag was first and last seen.
    \item The number of times it was detected \item the number of
    time it was detected multiple time in a frame
    \item an histogram of tracking gap for this tag. This histogram
    is only valid for experiment with a single space, and will
    report erroneous values for multiple space experiment.
    }
}
