% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Queries.R
\name{fmQueryIdentifyFrames}
\alias{fmQueryIdentifyFrames}
\title{Computes ant position during the experiment}
\usage{
fmQueryIdentifyFrames(
  experiment,
  start = NULL,
  end = NULL,
  computeZones = FALSE,
  singleThreaded = FALSE,
  showProgress = FALSE
)
}
\arguments{
\item{experiment}{the \code{\link{fmCExperiment}} to query for, use
\code{e$const} if you didn't opened the experiment in read
only mode.}

\item{start}{the starting time for the query. Data which were
acquire before that time will not be reported. Could be NULL
for -∞, a POSIXct a fmTime or fmTimeCPtr object.}

\item{end}{the starting time for the query. Data which were acquire
after that time will not be reported. Could be NULL for +∞, a
POSIXct a fmTime or fmTimeCPtr object/}

\item{computeZones}{reports the user defined zone for the
ants. otherwise the data.frame will be smaller and the
computation faster.}

\item{singleThreaded}{perform a single threaded query. Could be
faster for small experiment in number of Ant as identification
is mostly IO bounded and the overhead induced by
multi-threading synchronization impacts performances.}

\item{showProgress}{display the progress of the computation on the
standard error output. It may not be portable behavior on all
OS.}
}
\value{
a list of \linkS4class{fmIdentifiedFrame}, ordered in time.
}
\description{
Computes ant position during the experiment
}
