% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Space.R
\name{fmSpace}
\alias{fmSpace}
\title{a Reference Class wrapper for
    \href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Space/fort::myrmidon::Space}{fort::myrmidon::Space}}
\description{
a Reference Class wrapper for
    \href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Space/fort::myrmidon::Space}{fort::myrmidon::Space}
}
\section{Fields}{

\describe{
\item{\code{spaceID}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Space/fort::myrmidon::Space::SpaceID}{fort::myrmidon::Space::SpaceID}}

\item{\code{name}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Space/fort::myrmidon::Space::Name}{fort::myrmidon::Space::Name}}

\item{\code{setName}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Space/fort::myrmidon::Space::SetName}{fort::myrmidon::Space::SetName}}

\item{\code{createZone}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Space/fort::myrmidon::Space::CreateZone}{fort::myrmidon::Space::CreateZone}}

\item{\code{deleteZone}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Space/fort::myrmidon::Space::DeleteZone}{fort::myrmidon::Space::DeleteZone}}

\item{\code{cZones}}{returns a \linkS4class{fmCZoneByID} as
\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Space/fort::myrmidon::Space::CZones}{fort::myrmidon::Space::CZones}}

\item{\code{zones}}{returns a \linkS4class{fmZoneByID} as
\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Space/fort::myrmidon::Space::Zones}{fort::myrmidon::Space::Zones}}

\item{\code{locateMovieFrame}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Space/fort::myrmidon::Space::LocateMovieFrame}{fort::myrmidon::Space::LocateMovieFrame}}
}}

