% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone.R
\name{fmZone}
\alias{fmZone}
\title{A geometric zone of a space}
\description{
A fmZone represents a geometric zone of a
\link{\code{fmSpace}}.

Zone defines a named region of interest for tracking
and interactions. It means that two Ant, which are lying in two
separate Zone will never report a collision or Interaction.
}
\details{
Zones are uniquely idetinfied trough their ID in an
Experiment, but they are managed and are related to a single
\link{\code{fmSpace}}. They also have a user defined name, but
it is never used internally.

By default an Ant lies in no Zone, which is identified by
the ZoneID 0.

Zone have time valid ZoneDefinition which represents their
geometry. In most cases a Zone will have a single
ZoneDefinition valid for
]\link{\code{fmTimeSinceEver}};\link{\code{fmTimeForever}}[. However
it is possible to add as many different definition to a Zone,
as long as they do not overlap in Time. The defintions are
manipulated with \link{\code{fmZone$addDefinition}} and
\link{\code{DeleteDefinition}}.
}
\section{Fields}{

\describe{
\item{\code{name}}{the zone name}

\item{\code{ID}}{the zone ID, which is unique in the experiment}

\item{\code{definitions}}{(read-only) the zone definitions}
}}

\seealso{
Other fmZone classes and methods: 
\code{\link{fmZone$addDefinition}},
\code{\link{fmZone$deleteDefinition}},
\code{\link{fmZoneDefinition}}
}
\concept{fmZone classes and methods}
