% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zone.R
\name{fmZone}
\alias{fmZone}
\title{Reference Class wrapper for the c++
    \href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Zone/fort::myrmidon::Zone}
    class}
\description{
Reference Class wrapper for the c++
    \href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Zone/fort::myrmidon::Zone}
    class
}
\section{Fields}{

\describe{
\item{\code{addDefinition}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Zone/fort::myrmidon::Zone::AddDefinition}{fort::myrmidon::Zone::AddDefinition}}

\item{\code{cDefinitions}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Zone/fort::myrmidon::Zone::CDefinitions}{fort::myrmidon::Zone::CDefinitions}}

\item{\code{definitions}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Zone/fort::myrmidon::Zone::Definitions}{fort::myrmidon::Zone::Definitions}}

\item{\code{eraseDefinition}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Zone/fort::myrmidon::Zone::EraseDefinition}{fort::myrmidon::Zone::EraseDefinition}}

\item{\code{name}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Zone/fort::myrmidon::Zone::Name}{fort::myrmidon::Zone::Name}}

\item{\code{setName}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Zone/fort::myrmidon::Zone::SetName}{fort::myrmidon::Zone::SetName}}

\item{\code{zoneID}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::Zone/fort::myrmidon::Zone::ZoneID}{fort::myrmidon::Zone::ZoneID}}
}}

