% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone.R
\name{fmZoneDefinition}
\alias{fmZoneDefinition}
\title{Defines the geometry of a Zone for a given Time range}
\description{
Defines the geometry of a Zone for a given Time range.

fmZoneDefinition can only be created from a
\code{\link{fmZone}} with \code{\link{fmZone$addDefinition}()}
}
\section{Fields}{

\describe{
\item{\code{shapes}}{a \code{\link{fmShapeList}} defining the zone
geometry}

\item{\code{start}}{the first valid \code{\link{fmTime}} for the zone}

\item{\code{end}}{the last valid \code{\link{fmTime}} for the zone}
}}

\seealso{
Other fmZone classes and methods: 
\code{\link{fmZone$addDefinition}},
\code{\link{fmZone$deleteDefinition}},
\code{\link{fmZone}}
}
\concept{fmZone classes and methods}
