% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone.R
\name{fmZoneDefinition}
\alias{fmZoneDefinition}
\title{Defines the geometry of a Zone for a given Time range}
\description{
Defines the geometry of a Zone for a given Time range.
}
\section{Fields}{

\describe{
\item{\code{shapes}}{a \link{\code{fmShapeList}} defining the zone geometry}

\item{\code{start}}{the first valid time for the zone}

\item{\code{end}}{the last valid time for the zone}
}}

\seealso{
Other fmZone classes and methods: 
\code{\link{fmZone$addDefinition}},
\code{\link{fmZone$deleteDefinition}},
\code{\link{fmZone}}
}
\concept{fmZone classes and methods}
