% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zone.R
\name{fmZoneDefinition}
\alias{fmZoneDefinition}
\title{Reference Class wrapper for the C++
    \href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::ZoneDefinition/fort::myrmidon::ZoneDefinition}{fort::myrmidon::ZoneDefinition}
    class}
\description{
Reference Class wrapper for the C++
    \href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::ZoneDefinition/fort::myrmidon::ZoneDefinition}{fort::myrmidon::ZoneDefinition}
    class
}
\section{Fields}{

\describe{
\item{\code{geometry}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::ZoneDefinition/fort::myrmidon::ZoneDefinition::Geometry}{fort::myrmidon::ZoneDefinition::Geometry}}

\item{\code{setGeometry}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::ZoneDefinition/fort::myrmidon::ZoneDefinition::SetGeometry}{fort::myrmidon::ZoneDefinition::SetGeometry}}

\item{\code{start}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::ZoneDefinition/fort::myrmidon::ZoneDefinition::Start}{fort::myrmidon::ZoneDefinition::Start}}

\item{\code{end}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::ZoneDefinition/fort::myrmidon::ZoneDefinition::End}{fort::myrmidon::ZoneDefinition::End}}

\item{\code{setStart}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::ZoneDefinition/fort::myrmidon::ZoneDefinition::SetStart}{fort::myrmidon::ZoneDefinition::SetStart}}

\item{\code{setEnd}}{\href{https://formicidae-tracker.github.io/studio/docs/latest/api/#fort::myrmidon::ZoneDefinition/fort::myrmidon::ZoneDefinition::SetEnd}{fort::myrmidon::ZoneDefinition::SetEnd}}
}}

\examples{
\dontrun{
# sets the geomtery of a zone zd to two circle
zd$setGeometry(list(fmCircleCreate(c(0,0),3),fmCircleCreate(c(10,10,10))))
}
}
